GWASExperiment objects

The GWASExperiment class is a matrix-like container which directly inherits all the properties from the SummarizedExperiment class and extends it with respect to holding GWAS data in terms of accomodating not only genotypic and phenotypic data but also association scores and SNP effects from Genome-Wide Association tests. In addition, as PRISMA's main purpose is the construction and evaluation of Polygenic Risk Scores, it offers facilities to store the results of respective analyses.

Specifically, the assays of a GWASExperiment hold a matrix object (SnpMatrix or bigSNP) where rows represent SNPs or other variants and columns represent samples. The values of the assays represent genotypes and they are accessed and manipulated through the respective functions of the packages snpStats and bigSNP. Apart from the assays, the rowData of the GWASExperiment hold information about the markers and the colData hold information about the samples. Such information are typically derived from the initial input data used to construct the GWASExperiment. The easiest way to construct it is from a set of PLINK files in BED, BIM and FAM formats. Direct construction is also possible, as long as the provided information about markers and samples is complete, for example, markers should be accompanied at least by chromosome and position data as well as major/minor allele information.

Apart from the basic GWASExperiment data which make direct use of the SummarizedExperiment slots, the class offers slots to store associated phenotypic and anthropometric data for each sample which are used for GWA tests, and slots to store the results of these tests in terms of p-values and reggression coefficients (effects). For the SNPs that may then become part of a PRS, the GWASExperiment also offers a slot for storing the weights of the PRS SNPs. Furthermore, GWASExperiment takes advantage of the metadata slot of the SummarizedExperiment to store additional information on GWAS and subsequent PRS extraction through the GWA summary statistics, such as Principal Component Analysis results used as covariates to capture and account for potential underlying population stratification, reference genome information, track of the filters applied to the genotypes and samples and other information. Finally, it offers the ability to avoid genotypic and phenotypic data duplication when conducting multiple GWA tests by allowing the association and storage of many test results (e.g. with different covariates) within the same object.

Usage

## Constructor

    GWASExperiment(
        genotypes = SnpMatrix(), 
        features = DataFrame(),
        samples = DataFrame(),
        phenotypes = DataFrame(),
        pvalues = SimpleList(),
        effects = SimpleList(),
        prsbetas = SimpleList(),
        metadata = list(
            genome = NA_character_,
            backend = NA_character_,
            filters = setNames(
                data.frame(matrix(ncol = 5, nrow = 0)),
                c("parameter", "name", "value", "type", "filtered")
            )
        )
    )

    ## Accessors - apart from the ones of SummarizedExperiment

    genotypes(x, ...)
    genotypes(x, ...) <- value
    gfeatures(x, withDimnames=TRUE, ...)
    gfeatures(x, ...) <- value
    gsamples(x, withDimnames=TRUE, ...)
    gsamples(x, ...) <- value

    phenotypes(x, withDimnames=TRUE, ...)
    phenotypes(x, withDimnames=TRUE, ...) <- value

    pvalues(x, response = 1L, covariates = NULL, npcs = 0,
        withDimnames = TRUE, ...)
    pvalues(x, response = 1L, covariates = NULL, npcs = 0,
        withDimnames = TRUE, ...) <- value
    allPvalues(x, i = seq_len(ncol(x)), ...)
    effects(x, response = 1L, covariates = NULL, npcs = 0,
        withDimnames = TRUE, ...)
    effects(x, response = 1L, covariates = NULL, npcs = 0,
        withDimnames = TRUE, ...) <- value
    allEffects(x, i = seq_len(ncol(x)), ...)
    prsbetas(x, response = 1L, covariates = NULL, npcs = 0,
        withDimnames = TRUE, ...)
    prsbetas(x, response = 1L, covariates = NULL, npcs = 0,
        withDimnames = TRUE, ...) <- value
    allPrsbetas(x, i = seq_len(ncol(x)), ...)
    
    
    ## Splitting
    
    gsplit(x, by, across = c("features", "samples"), 
        rc = NULL, ...)

    ## Elements
    
    genome(x)
    genome(x, ...) <- value
    gdsfile(x)
    gdsfile(x, ...) <- value
    filterRecord(x)
    filterRecord(x, ...) <- value
    pcaCovariates(x)
    pcaCovariates(x, ...) <- value

    ## Quick colData access

    # S4 method for GWASExperiment,ANY,missing
[[(x, i, j, ...)
    # S4 method for GWASExperiment,ANY,missing
[[(x, i, j, ...) <- value

    ## Subsetting

    # S4 method for GWASExperiment
[(x, i, j, ..., drop=TRUE)
    # S4 method for GWASExperiment,ANY,ANY,GWASExperiment
[(x, i, j) <- value

    ## Combining

    # S4 method for GWASExperiment
cbind(..., deparse.level=1)
    # S4 method for GWASExperiment
rbind(..., deparse.level=1)

Arguments

x

a GWASExperiment object.

...

for cbind, rbind, ... contains GWASExperiment objects to be combined. For other accessors, ignored.

value

an object of a class specified in the S4 method signature or as outlined in Details.

i, j

for genotypes, genotypes<-, i is an integer or numeric scalar, see Details for additional constraints and for i in functions like pvalues. Brifely, for the latter it is a scalar or vector of indices for which to retrieve statistical scores, corresponding to a number of models fit to x.

For [,GWASExperiment, [,GWASExperiment<-, i, j are subscripts that can act to subset the rows and columns of x, that is the matrix elements of genotypes.

For [[,GWASExperiment, [[<-,GWASExperiment, i is a scalar index (e.g., character(1) or integer(1)) into a column of gsamples.

withDimnames

a logical(1), indicating whether the dimnames of the GWASExperiment object should be applied (i.e. copied) to the extracted assays (genotypes or other data such as samples, features, phenotypes). More precisely, setting withDimnames=FALSE in the getter returns the assays as-is whereas setting withDimnames=FALSE return them with possibly modified dimnames.

Setting withDimnames=FALSE in the setter (assays<-) is required when the dimnames on the supplied assays are not identical to the dimnames on the GWASExperiment object. It does not influence actual assignment of dimnames to assays (they're always stored as-is). For more details, see SummarizedExperiment.

drop

A logical(1), ignored by these methods.

deparse.level

See ?base::cbind for a description of this argument.

genotypes

a SnpMatrix or bigSNP matrix-like object that holds the genotypes for the SNPs/markers in each row and each sample in each column.

features

a DataFrame or data.frame holding detailed information about each genotyped SNP. See also Details.

samples

a DataFrame or data.frame holding information about each genotyped sample. See also Details.

phenotypes

a DataFrame or data.frame holding phenotypic records about each genotyped sample, typically used for GWA tests. See also Details.

pvalues

a SimpleList or list with p-values from GWA tests for each SNP. Each member corresponds to a GWA test with specific covariates. See also Details.

effects

a SimpleList or list with regression coefficients (or converted odds ratios) from GWA tests for each SNP (effects). Each member corresponds to a GWA test with specific covariates. See also Details.

prsbetas

a SimpleList or list with weights for each SNP participating in a PRS. Each member corresponds to a GWA test with specific covariates. Typically, subset of effects, unless an adjustment algorithms (e.g. lassosum) has been applied. See also Details.

metadata

a list similar to metadata of SummarizedExperiment. It should contain at least genome, backend and filters members. See also Details.

response

a numeric or integer or character scalar. If a number, it should correspond to the column number of the response phenotype in a GWA test. If character, it should be one of the phenotypes associated with x. See also Details.

covariates

a character vector or NULL. The covariates to be used in regression models and GWA tests. They should exist in the phenotypes of x. See also Details.

npcs

the number of PCs included in the covariates in a GWA test. See also Details.

by

the name of the sample or SNP attribute to be used as the splitting factor.

across

split a GWASExperiment accross "features" (default) or "samples", according to the attribute defined by by.

rc

fraction of available cores to use for splitting.

Details

The GWASExperiment class is a straighforward expansion of the SummarizedExperiment class which is meant for numeric and other data types derived typically from a sequencing experiment but is suitable to host genotypic data as the main features of the latter are Single Nucleotide Polymorphisms (SNPs) which share a lot of common information with sequencing experiments such as genomic coordinates and certain types of abundance. The structure is rectangular like a matrix, but with additional annotations on the rows and columns, and with the possibility to manage several assays simultaneously so long as they be of the same dimensions. More generic information can be found on the manual pages of the SummarizedExperiment class. The following focus on the expansions introduced in the GWASExperiment class.

The rows of a GWASExperiment object represent SNPs/markers of interest. Information about these markers is stored in a DataFrame or data.frame object, accessible using the function gfeatures (or the native rowData as gfeatures is an alias to it). The object must have as many rows as there are rows in the GWASExperiment object, with each row of the object providing information on the SNP in the corresponding row of the GWASExperiment object. Columns of the object represent different attributes of the features of interest. Specifically, for the GWASExperiment case, it must have at least genomic coordinates, that is chromosome and position, a unique id (it can be a dbSNP id) and the two alleles comprising the genotype. A typical case is when the GWASExperiment is created from PLINK files with the backend provided by snpStats package and the columns of the object are chromosome, snp.name, cM, position, allele.1 and allele.2. These are also the preferred column names of the object. It must also have assigned rownames, preferrably the same as the snp.names column.

Each column of a GWASExperiment object represents a sample (e.g. a person contributing to a GWAS project). Information about the samples are stored in a DataFrame or , data.frame object, accessible using the function gsamples (or the native colData as gsamples is an alias to it). The object must have as many rows as there are columns in the GWASExperiment object, with each row of the object providing information on the sample in the corresponding column of the GWASExperiment object. Columns of the object represent different sample attributes, e.g., pedigree, sex etc. Column names typically provide a short identifier unique to each sample. A typical case is when the GWASExperiment is created from PLINK files with the backend provided by snpStats package and the columns of the object are pedigree, member, father, mother, sex and affected. These are also the preferred column names of the object. It must also have assigned rownames, preferrably the same as the member column.

A GWASExperiment object, as a SummarizedExperiment, can also contain information about the overall experiment, for instance the lab in which it was conducted, the publications with which it is associated, etc. This information is stored as a list object, accessible using the metadata function. The form of the data associated with the experiment is left to the discretion of the user. However, the metadata of a GWASExperiment object must have at least three elements. These are:

genome: the reference genome used in the genotyping technology (usually hg38 or hg19). It can be accessed and set using the genome function.

filters: the set of filters applied to the GWASExperiment to filter out noisy genotyping data (if any) and a summary of their results in a data.frame. A placeholder is automatically added on construction. It can be accessed and set with the filterRecord fuction.

backend: the backend package used to read and store the genotype data (one of snpStats or bigsnpr). This is crucial for downstream functions and manipulation and must not be changed.

The GWASExperiment container is appropriate for matrix-like genotypic data, such as those derived from gentype microarray technologies. Apart from the related SummarizedExperiment data accessor functions, the data are accessed using the genotypes function, described below. This returns a SnpMatrix or bigSNP object. A SimpleList object with genotypic data from potentially different experiment can be accessed with assays. genotypes returns the first one, but typically, this is what is being used. Each element of the list must itself either a SnpMatrix or a codebigSNP object and must have dimensions that are the same as the dimensions of the GWASExperiment in which they are stored. Row and column names of each object must either be NULL or match those of the GWASExperiment during construction.

Furthermore, the GWASExperiment offers facilities for storing other data typical for subsequent GWAS analyses and PRS constructions, such as storing phenotypic and anthropometric data used for association analysis, the results of GWA tests and the results of PRS analysis. Instead of storing these data to the metadata slot of a SummarizedExperiment, they are bound to slots following the dimensions and subsetting of the main genotypic matrix object. Therefore, phenotypic data are accessed through the phenotypes function, the results of GWA tests (p-values, regression coefficients) are accessed through the pvalues and effects functions respectively, and associated PRS weights are accessed throught the function prsbetas. All these functions are described below and all return SimpleList objects, as a GWASExperiment object may have multiple GWA and PRS results attached to it to avoid genotype repetition. A listing of associated analyses with a GWASExperiment object may be retrieved with the listGwa and listPrs functions and the functions allPvalues, allEffects and allPrsbetas can be used to retrieve a list of all the analyses results.

Constructor

GWASExperiment instances are constructed using the GWASExperiment function with arguments outlined above.

Accessors

In the following code snippets, x is a GWASExperiment object. We do not describe standard SummarizedExperiment accessors such as rowData, colData, dim etc. Details can be found in the respective man pages. General patterns such as that row data must have NULL or identical row names as with the main matrix object (here, the genotypes) are assumed.

genotypes(x), genotypes(x) <- value:

Get or set the genotypes. value is a SnpMatrix or bigSNP with the same dimensions as x.

gfeatures(x, withDimnames=TRUE), gfeatures(x) <- value:

Get or set the feature data (similar to row data in a SummarizedExperiment). value is a DataFrame or a data.frame object.

gsamples(x, withDimnames=TRUE), gsamples(x) <- value:

Get or set the sample data (similar to column data in a SummarizedExperiment). value is a DataFrame or a data.frame object.

phenotypes(x, withDimnames=TRUE), phenotypes(x) <- value:

Get or set the phenotypic data for each sample. value is a data.frame object with as many rows as the number of samples and associated phenotypes in the experiment.

pvalues(x, response = 1L, covariates = NULL, npcs = 0, withDimnames = TRUE), pvalues(x, response = 1L, covariates = NULL, npcs = 0, withDimnames = TRUE, ...) <- value:

Get or set the p-values after the application of one or more genotype-phenotype association tests. response is the phenotype for which the association test is performed. It can be a numeric or integer value denoting the column number in the phenotypes data frame which is associated with x value, or a character in which case it's the name of the phenotype and corresponds to a column name of the aforementioned data frame. It defaults to the first available phenotype. covariates are the phenotypic covariates to be included in an association model and must be available in the phenotypes data frame. It can be either a numeric/integer vector denoting column numbers or a character vector denoting column names of the phenotypes data frame. It defaults to NULL for no addition of covariates, other than the main phenotype (response). npcs denotes the number of PCs to include in the model as additional covariates to capture potential population substructures. It defaults to 0 for no PCs as additional covariates. For withDimnames, see above e.g. gfeatures. The reason for this accessor to have additional arguments (response, covariates, npcs) is that x may have multiple association test results attached to it, and these variable offer a way of unique access to them. Finally, value is a numeric matrix with p-values for each SNP. This matrix has as many columns as the different test types performed (e.g. SNPTEST and statgenGWAS) and are named accordingly.

allPvalues(x, i = seq_len(ncol(x))):

This will retrieve (access-only) either all the p-value matrices attached to x or as many as requested by i, assuming at least i tests have been performed.

effects(x, response = 1L, covariates = NULL, npcs = 0, withDimnames = TRUE), effects(x, response = 1L, covariates = NULL, npcs = 0, withDimnames = TRUE, ...) <- value:

Get or set the SNP effects with respect to the phenotype after the application of one or more genotype-phenotype association tests. The effects are essentially regression coefficients or odds ratios for binomial regression models. response, covariates and npcs are as in the pvalues accessor. value is a numeric matrix with effects for each SNP. This matrix has as many columns as the different test types performed (e.g. SNPTEST and statgenGWAS) and are named accordingly.

allEffects(x, i = seq_len(ncol(x))):

This will retrieve (access-only) either all the effect matrices attached to x or as many as requested by i, assuming at least i tests have been performed.

prsbetas(x, response = 1L, covariates = NULL, npcs = 0, withDimnames = TRUE), prsbetas(x, response = 1L, covariates = NULL, npcs = 0, withDimnames = TRUE, ...) <- value:

Get or set the beta coefficients (or odd ratios) effects derived after the application of a Polygenic Risk Scores algorithm, with respect to the phenotype tested with one or more of the association tests. Essentially, the PRS betas are the regression coefficients for the SNPs participating in a PRS either corrected in the case of effect shrinkage algorithms (e.g. lassosum), or as returned by the association test. response, covariates and npcs are as in the pvalues accessor. value is a numeric matrix with effects for each SNP participating in a PRS candidate and 0 otherwise. This matrix has as many columns as the different PRS algorithms executed on x and are named accordingly.

allPrsbetas(x, i = seq_len(ncol(x))):

This will retrieve (access-only) either all the PRS beta matrices attached to x or as many as requested by i, assuming at least i tests have been performed.

genome(x), genome(x) <- value:

Get or set the genome build where the features (SNPs, other variations) are mapped to. Typically, one of "hg19" or "hg38" as the class is designed mainly for human data. Other genomes may be provided, but for these human builts, other functionalities are also supported such as SNP information retrieval from dbSNP and/or Ensembl, and also enrichment of PGS score files.

gdsfile(x), gdsfile(x) <- value:

Get or set the supporting GDS file location. This is required by the package SNPRelate to perform certain functionalities such as Linkage Disequilibrium (LD) pruning, Identity By Descent (IBD) filtering and others. If not provided, these functionalities will not be available.

filterRecord(x), filterRecord(x) <- value:

Get or set a data frame holding a record of the filters applied to x, the cutoff values and the filtering results (i.e. how many SNPs where filtered). This data frame has a specific format and violating will cause an error. Typically, users will only access the filter records. To see the structure of this data frame, call filterRecord(x).

pcaCovariates(x), pcaCovariates(x) <- value:

Get or set the PCA covariates that capture underlying population stratification structures. The PCA covariates are typically calculated using the calcPcaCovar function with one of the supporting methods within the package. Users will usually only access the PCA covariates while setting them using the calcPcaCovar function.

gsplit(x, by, across = c("features", "samples"), rc = NULL:

Split an object by column names stored in the gfeatures or gsamples. If the former, the object is split row-wise while if the former, it is split column-wise. by is either the name of the gfeatures or gsamples column, or its index. If by is missing, then the object is split according to the rc variable which prepares the object for parallel calculatios and the split is done according to the number of available cores to use (rc). If rc=NULL and also by is missing, then there is no split and the object is returned in a lenght-one list.

Subsetting

The subsetting operations are as described in the respective section of the SummarizedExperiment help page. There are restrictions in the case of a GWASExperiment as in the case of replacement, the data types stored in its slots must be compatible (i.e. SnpMatrix, bigSNP objects for genotypes and DataFrame, data.frame objects for associated genotype and sample data).

Combining

The combining functions of GWASExperiment objects follow the same principles as those of SummarizedExperiment objects. Special attention is given regarding the integrity of the main SummarizedExperiment class extensions. These are briefly the phenotypes which follow the samples (columns/colData) of the GWASExperiment object and statistical scores parallel to the features (rowData) of the object (pvalues, effects etc.).

When combining objects, the following attributes and values are not preserved, as they become meaningless with the combination of samples and/or SNPs:

pvalues, effects, prsbetas: these scores are dropped as the initial population in potentially both samples and SNPs from which they were derived is changed.

filters: any filtered SNPs/samples statistics are re-initialized as the population of SNPs/samples is changed.

any values related to PCA, as again, these statistics depend on the population which changes with object combination.

Splitting

The splitting operations differ from the classic SummarizedExperiment in the sense that a GWASExperiment object may curry multiple associated GWA and PRS analyses using different sets of covariates. These can be preserved by using the gsplit function which operates across feature properties or sample properties.

Author

Panagiotis Moulos

See also

Examples

library(PhenotypeSimulator)

# Simulate some data
dat1 <- makeSimData(nsnp=100,nsam=10,nphe=3,csnp=10)
#> coercing object of mode  numeric  to SnpMatrix

# Simple GWASExperiment object
gwe1 <- GWASExperiment(
    genotypes=dat1$snp,
    features=dat1$feature,
    samples=dat1$sample,
    phenotypes=dat1$pheno
)

# With pseudo-effects and pseudo-pvalues
pspval <- matrix(runif(200),100,2)
rownames(pspval) <- rownames(dat1$snp)
colnames(pspval) <- c("test1","test2")

# Pseudo-effects and directions
pseff <- matrix(runif(200,min=0,max=2),100,2)
dire <- matrix(FALSE,100,2)
dire[sample(nrow(dire),50),1] <- TRUE
dire[sample(nrow(dire),50),2] <- TRUE
pseff[dire[,1],1] <- -pseff[dire[,1],1]
pseff[dire[,2],2] <- -pseff[dire[,2],2]
rownames(pseff) <- rownames(dat1$snp)
colnames(pseff) <- c("test1","test2")

# Assign effects and p-values to the object, assuming a model
# with 'Trait_1' as response and 'Trait_2', 'Trait_3' as
# covariates has been fit
pvalues(gwe1,response="Trait_1",
    covariates=c("Trait_2","Trait_3")) <- pspval
effects(gwe1,response="Trait_1",
    covariates=c("Trait_2","Trait_3")) <- pseff

# Add SNPs (rbind)
dat2 <- makeSimData(nsnp=100,nsam=10,nphe=3,csnp=10)
#> coercing object of mode  numeric  to SnpMatrix
rownames(dat2$snp) <- rownames(dat2$feature) <- 
    paste("snp",101:200,sep="_")
gwe2 <- GWASExperiment(
    genotypes=dat2$snp,
    features=dat2$feature,
    samples=dat1$sample,
    phenotypes=dat1$pheno
)
# p-values and effects will be dropped with warning
gwe12 <- suppressWarnings(rbind(gwe1,gwe2))

# Add samples (cbind)
dat3 <- makeSimData(nsnp=100,nsam=10,nphe=3,csnp=10)
#> coercing object of mode  numeric  to SnpMatrix
rownames(dat3$sample) <- rownames(dat3$pheno) <- 
    colnames(dat3$snp) <- paste("sample",11:20,sep="_")
gwe3 <- GWASExperiment(
    genotypes=dat3$snp,
    features=dat1$feature,
    samples=dat3$sample,
    phenotypes=dat3$pheno
)
# p-values and effects will be dropped with warning
gwe13 <- suppressWarnings(cbind(gwe1,gwe3))