PRS weight adjustment through final regression

This function may be used to adjust the SNP effects to derive the final PRS (e.g. with the total dataset) after the whole extraction and assessment process. The process is not obligatory.

Usage

adjustPrsWeights(snpSelection, gwe, response, 
        covariates = NULL, pcs = FALSE, family = NULL,
        rc = NULL)

Arguments

snpSelection

A data frame with SNPs comprising a PRS, properly formatted (see aggregatePrsMarkers).

gwe

The GWASExperiment object used to derive the PRS.

response

A character with the phenotypic response used to derive the PRS.

covariates

A character vector with covariares used to derive the PRS.

pcs

Use the PCs present in gwe? Defaults to FALSE. If the PRS has been extracted with PCs as covariates, it should be TRUE.

family

Regression family, e.g. "gaussian", see also gwa.

rc

The fraction of available cores to use for parallel calculations. Default is NULL for no parallelization.

Value

A new data frame with SNPs comprising the PRS with adjusted effects.

Author

Panagiotis Moulos

Examples

    data(toy,package="prisma")
    data(prisma_out,package="prisma")
    
    prs <- getPrsCandidates(prismaOut,"statgen",1)
    #prs <- prismaOut$results$statgen$candidates[[1]]
    
    prsa <- adjustPrsWeights(prs,toy,"BMI",
        c("Age","Gender","Case_Control"),pcs=FALSE)
#> 
#> Performing GWAS with GLM over 1 chunks
#> Trait(s)    : BMI
#> Covariate(s): Age, Gender, Case_Control
#> Regression  : gaussian (predicted)
#> Use PCs     : No
#>   testing SNPs from rs1442343 to rs6506142
#> Found 3 associations (uncorrected p-values).
    
    identical(prs$effect_weight,prsa$effect_weight)
#> [1] FALSE
    # FALSE