Aggregate PRS candidate SNPs from iterative extraction

This function aggregates the results of an iterative PRS extraction derived using prsPipeline. The aggregation is done by intersecting or unifying the SNPs extracted by more than one PRS algorithms. The aggregated result can be then subjected to a PRS selection process according to the frequency of appearance of each marker and result in a subset of the aggregated markers as the final PRS.

Usage

aggregatePrsMarkers(gwaList, mode = c("intersect", 
        "union"), qcut = 0.9, assoc = c("auto", "glm",
        "rrblup", "statgen", "snptest", "plink", "lasso"),
        avgfun = c("mean", "median", "weight"), gwe = NULL)

Arguments

gwaList

A list of GWASExperiments or their summaries as derived from prsPipeline.

mode

Aggregation mode, can be "intersect" (default) or "union".

qcut

A quantile cutoff for the frequency of SNP appearances in gwaList. SNPs with frequencies below the one corresponding to this quantile are ignored from the output.

assoc

Which effects to use out of the performed GWA tests as the SNP weights in PRS. Can be "auto" (default) or any of "glm", "rrblup", "statgen", "snptest", "plink", as long as it exists. See also Details.

avgfun

How to summarize SNP effects (regression coefficients) after the iterative extraction process. One of "mean" (default), "median" or "weight". See also Details.

gwe

The GWASExperiment object used to derive the PRS. If given, the output is a more complete data frame, harmonized with the scores from EBI PGS Catalog.

Value

If gwe is not NULL, a data frame of SNP markers summarized a prsPipeline run with the following columns:

chromosome: the chromosome where the marker is located

position: the marker position in chromosome

variant_id: SNP rs id or other identification where available

risk_allele: The risk allele

reference_allele: The non-risk allele

locus_name: The SNP's closest gene (NA here, assigned later through assignGenes)

effect_weight: The SNP effect as derived from regression functions and avgfun

OR: The effect_weight transformed to Odds Ratio

asm: The genome assembly ("hg19" or "hg38" or "nr" if unknown)

freq: The number of appearances in the prsPipeline iterative process

If gwe is NULL, a simpler data frame of SNP marker with the following columns:

snp SNP rs id or other identification where available

freq The number of appearances in the prsPipeline iterative process

effect The SNP effect as derived from regression functions and avgfun

Details

Regarding the assoc argument, this controls which effects should be included in the final PRS data frame, out of the available ones. If "auto" (the default), the final effects are selected based on the GWA algorithms that have been executed and an internal prioritization scheme according to our perception regarding algorithm popularity. Specifically, the first available effects are used based on the following prioritization:

SNPTEST ("snptest")

PLINK ("plink")

statgenGWAS ("statgen")

GLM ("glm")

rrBLUP ("rrblup")

In the future we intend to combine all available effects based on linear optimization methods (work in progress)

Regarding the avgfun argument, this controls how the SNP effects in PRS candidates are summarized after the iterative PRS extraction process in order to produce the final effects for the calculation of weighted PRS. When "mean" (default), the final effect for a SNP is the mean off all its effects across the iterations and when it appears. When "median", it is the respective median while when "weight", the SNP effects are weighted by 1 minus the recorded PRS R^2 so that the "best" runs can be given more value.

Author

Panagiotis Moulos

Examples

    data(toy,package="prisma")
    data(prisma_out,package="prisma")
    
    iters <- prismaOut$results$glm$iterations
    #iters <- getRawIterations(prismaOut,"glm")
    snps <- aggregatePrsMarkers(iters,qcut=0.25,gwe=toy)
#> Aggregation mode: intersect
#>   iteration 1 yields 3 markers
#>   iteration 2 yields 62 markers
#>   iteration 3 yields 40 markers
#>   iteration 4 yields 2 markers
#> Calculating frequencies
#> Averaging effects with mean
#> Constructing output
#> Done!