Calculate PCA covariates for GWASExperiments

This function performs PCA using three available methods to the genetic data included in the iput object in order to capture potential underlying population structure or outliers, and derive PCs to be used as covariates in downstream GWA. Two of the three methods can be used to perform robust PCA with the intention of better capturing potential outliers.

Usage

calcPcaCovar(obj, ld = 0.2, method = c("auto", 
        "snprel", "grid", "hubert"), npc = 0, pval = 0.05, 
        rc = NULL)

Arguments

obj

A GWASExperiment object without missing values.

ld

Linkage Disequilibrium cutoff (defaults to 0.2) for SNPs to include in PCA. See also Details.

method

Method used to calculate PCs. One of "auto" (default), "snprel", "grid" or "hubert". See also Details.

npc

The number of PCs to attach to obj. When 0 (default), the Tracy-Widom test is used to determine the best number of PCs to include.

pval

p-value for the Tracy-Widom test. Defaults to 0.05. PCs with p-value below this threshold are ignored.

rc

Fraction of available cores to use when method="snprel".

Value

The input GWASExperiment object with attached PC covariates corresponding to the data hosted in the object. They can be retrieved with pcaCovariates(obj).

Details

Regarding the ld argument, it is generally recommended in the literature to use LD-pruned SNPs for PCA in GWAS Experiments. This argument controls the threshold of LD for LD-pruning. It is based on recommendations from Reed at al., 2013.

Regarding the method argument, when "auto", if PCs already exist in the object, LD pruning is redone following by robust PCA if the attached filters in the object have robust PCA marked. If not, PCA with the package SNPRelate is performed. It should be noted that for larger datasets, robust PCA can take a lot of time, while much less for the optimized method in SNPRelate.

Author

Panagiotis Moulos

Examples

    data(toy,package="prisma")
    gdsfile(toy) <- system.file(package="prisma","extdata","toy.gds")
    
    # Only 1 PC will be returned because of dataset size
    toy <- suppressWarnings(calcPcaCovar(toy,method="snprel"))
#> Performing PCA to select covariates to include in associations
#> Reading GDS file /tmp/RtmplLIKfg/temp_libpath4f71185f9e1c/prisma/extdata/toy.gds...
#> Performing LD pruning...
#> SNP pruning based on LD:
#> Excluding 0 SNP (non-autosomes or non-selection)
#> Excluding 38 SNPs (monomorphic: TRUE, MAF: NaN, missing rate: NaN)
#>     # of samples: 100
#>     # of SNPs: 962
#>     using 1 thread
#>     sliding window: 500,000 basepairs, Inf SNPs
#>     |LD| threshold: 0.2
#>     method: composite
#> Chromosome 18: 23.44%, 154/657
#> Chromosome 21: 21.28%, 73/343
#> 227 markers are selected in total.
#> LD pruning finished! 227 SNPs will be used for PCA.
#> 
#> Performing PCA...
#> Principal Component Analysis (PCA) on genotypes:
#> Excluding 773 SNPs (non-autosomes or non-selection)
#> Excluding 0 SNP (monomorphic: TRUE, MAF: NaN, missing rate: NaN)
#>     # of samples: 100
#>     # of SNPs: 227
#>     using 1 thread
#>     # of principal components: 32
#> PCA:    the sum of all selected genotypes (0,1,2) = 10268
#> CPU capabilities: Double-Precision SSE2
#> Fri Jun  3 14:28:17 2022    (internal increment: 13024)
#> [..................................................]  0%, ETC: ---        [==================================================] 100%, completed, 0s
#> Fri Jun  3 14:28:17 2022    Begin (eigenvalues and eigenvectors)
#> Fri Jun  3 14:28:17 2022    Done.
#> Selecting number of PCs to include
#>   with Tracy-Widom test for significant eigenvalues
#>     selected 1 PCs