Functions to summarize PRS cross-validations

These functions facilitate the summarization of lenghty PRS candidate cross-validation runs using multiple linear regression samples agains the main phenotype under investigation. They are not mean to be used directly (although not prohibited if you know what to expect and how to interpret) but they are rather used to generate plots and numbers for the PRISMA final HTML report.

Usage

summarizeCvMetrics(cvm)
    
    cvSummary(M, nsnp, scr = TRUE, out = FALSE)
    
    plotCvMetrics(cvo, what = c("r2", "rmse", "mae", "pr2","crl"),
        pl = c("pct", "prs"), silent = FALSE)

Arguments

cvm

the output of link{prismaCrossValidate} for one GWA method. See also Details.

M

a list member of cvm above for one PRS candidate, for one leave-out fraction. Contains metrics at this particular leave-out percentage of total samples. See examples.

nsnp

number of SNPs in PRS candidate. If not provided, guessed from M.

scr

print the output of cvSummary to screen? See also Details.

out

store the output of cvSummary to a variable? See also Details.

cvo

same as cvm.

what

which metric to plot out of the available ones? See also Details.

pl

"pct" for plots according to leave-out samples for all PRS candidates, "prs" for plots according to numbr of SNPs in PRS for a given percentage. Essentially, the two types of CV plots in a PRISMA report.

silent

Do the plotting without messages? Defaults to FALSE.

Value

For summarizeCvMetrics, a list of length the same as the PRS candidates for a specific GWA method. Each list member contains the metrics collected by prismaCrossValidate across the requested number of iterations. See also the man page of prismaCrossValidate. For cvSummary, the summary of CV metrics for a PRS candidate across the requested iterations in prismaCrossValidate. See Details for a detailed explanation of what is calculated. For plotCvMetrics, a list of the requested ggplot2 objects, named according to the type of plots (leave-out percentage based, or PRS size based). See also examples.

Details

Reagrding the cvm argument, this is a list of lists, containing several cross-validation metrics. The main members are named the number of SNPs in a PRS candidate and the sub-members are named after the leave-out splits of the cross validation. These structures are not meant to be used directly, although provided. They are to feed the overall PRISMA HTML report. Typically, this is a member of the list output produced by the function prismaCrossValidate which has as many cvm's as the number of GWA methods that were used in the main prisma run.

Regarding the scr and out variables, these control how should the summarized cross-validation metrics should be displayed and/or stored. If out is TRUE, the output is stored to variable for later use e.g. from prismaReport while if scr is TRUE the output is nicely printed on screen. If both are TRUE, then also a varable is returned. In this case be careful of the syntax as the screen may be flooded with information! Both cannot be set to FALSE. You will need some output!

Regarding the what argument, the following explain the different options:

r2: model R^2 values (train samples)

rmse: Root Mean Squared Error

mae: Mean Absolute Error

pr2: predicted R^2 values (LOO samples)

crl: correlations with phenotype

Regarding the output of cvSummary, the following are calculated for a PRS candidate, for a specific percentage of leave-out samples and for a specific number of cross-validation iterations:

General information

Number of SNPs in PRS (n_snp): The number of SNPs in the PRS candidate being evaluated.

Number of cross-validation (cv_times): The number of leave-out cross-validations. In each iteration, a random number of samples comprising the requested percentage of leave-out samples of the dataset is left out and the PRS validated based on regressing the phenotypes of the rest samples including the PRS as a covariate and evaluating the model based on the left-out samples.

Statistical significance of the PRS in model (max_prs_pvalue): The best p-value of the PRS as a covariate in the CV regression model.

R^2 summary statistics

R^2 of the full regression model including the PRS (mean_full_r2, sd_full_r2): The average R^2 of the full model including the PRS +/- standard deviation.

R^2 of the reduced regression model excluding the PRS (mean_reduced_r2, sd_reduced_r2): The average R^2 of the reduced regression model, including all other covariates but excluding the PRS, +/- standard deviation.

Adjusted R^2 of the PRS full model contribution (mean_prs_r2, sd_prs_r2): The average adjusted R^2 of the full model, denoting the contribution of the PRS (essentially R^2 of the full model minus the R^2 of the reduced model), +/- standard deviation.

Statistical significance of full R^2 against reduced R^2

p-value of a Student's t-test (ttest_full_reduced_r2) for the hypothesis of the mean R^2 of the full model being larger than the mean R^2 of the reduced model.

p-value of a Wiloxon rank sum test (wilcox_full_reduced_r2) for the hypothesis of the median R^2 of the full model being larger than the median R^2 of the reduced model.

Empirical p-value (empirical_full_reduced_r2) denoting the number of times the R^2 of the full model was smaller than the R^2 of the reduced model across the number of cross-validation iterations.

RMSE (Root Mean Squared Error) summary statistics

RMSE of the full regression model including the PRS (mean_full_rmse, sd_full_rmse): Average RMSE of the full model with the PRS +/- standard deviation.

RMSE of the reduced regression model excluding the PRS (mean_reduced_rmse, sd_reduced_rmse): Average RMSE of the reduced model including all other covariates but excluding the PRS, +/- standard deviation.

Statistical significance of full RMSE against reduced RMSE

p-value of a Student's t-test (ttest_full_reduced_rmse) for the hypothesis of the mean RMSE of the full model being larger than the mean RMSE of the reduced model.

p-value of a Wiloxon rank sum test (wilcox_full_reduced_rmse) for the hypothesis of the median RMSE of the full model being larger than the median RMSE of the reduced model.

Empirical p-value (empirical_full_reduced_rmse) denoting the number of times the RMSE of the full model was smaller than the RMSE of the reduced model across the number of cross-validation iterations.

MAE (Mean Absolute Error) summary statistics

MAE of the full regression model including the PRS (mean_full_mae, sd_full_mae): Average MAE of the full model with the PRS +/- standard deviation.

MAE of the reduced regression model excluding the PRS (mean_reduced_mae, sd_reduced_mae): Average MAE of the reduced model including all other covariates but excluding the PRS, +/- standard deviation.

Statistical significance of full MAE against reduced MAE

p-value of a Student's t-test (ttest_full_reduced_mae) for the hypothesis of the mean MAE of the full model being larger than the mean MAE of the reduced model.

p-value of a Wiloxon rank sum test (wilcox_full_reduced_mae) for the hypothesis of the median MAE of the full model being larger than the median MAE of the reduced model.

Empirical p-value (empirical_full_reduced_mae) denoting the number of times the MAE of the full model was smaller than the MAE of the reduced model across the number of cross-validation iterations.

Correlation summary statistics

R between observed & predicted values with the full model (mean_full_pred_cor, sd_full_pred_cor): The average correlation between the actual observed phenotype values and the predicted by the full regression model, including the PRS, +/- standard deviation.

R between observed & predicted values with the reduced model (mean_reduced_pred_cor, sd_reduced_pred_cor): The average correlation between the actual observed phenotype values and the predicted by the reduced regression model, including all the other covariates but excluding the PRS, +/- standard deviation.

Statistical significance of full R against reduced R

p-value of a Student's t-test (ttest_full_reduced_pred_cor) for the hypothesis of the mean R of the full model being larger than the mean R of the reduced model.

p-value of a Wiloxon rank sum test (wilcox_full_reduced_pred_cor) for the hypothesis of the median R of the full model being larger than the median R of the reduced model.

Empirical p-value (empirical_full_reduced_pred_cor) denoting the number of times the R of the full model was smaller than the R of the reduced model across the number of cross-validation iterations.

R^2 summary statistics for the predicted test values

R^2 of the full regression model including the PRS (mean_full_pred_r2, sd_full_pred_r2): The average R^2 of the full model including the PRS +/- standard deviation, for the predicted leave-out test values.

R^2 from the reduced regression model excluding the PRS (mean_reduced_pred_r2, sd_reduced_pred_r2): The average R^2 of the reduced regression model, including all other covariates but excluding the PRS, +/- standard deviation, for the predicted leave-out test values.

Adjusted R^2 from the PRS full model contribution (mean_prs_pred_r2, sd_prs_pred_r2): The average adjusted R^2 of the full model, denoting the contribution of the PRS (essentially R^2 of the full model minus the R^2 of the reduced model), +/- standard deviation, for the predicted leave-out test values.

Significance of full R^2 against reduced R^2 (predicted)

p-value of a Student's t-test (ttest_full_reduced_pred_r2) for the hypothesis of the mean R^2 of the full model being larger than the mean R^2 of the reduced model, for the predicted test values.

p-value of a Wiloxon rank sum test (wilcox_full_reduced_pred_r2) for the hypothesis of the median R^2 of the full model being larger than the median R^2 of the reduced model, for the predicted test values.

Empirical p-value (empirical_full_reduced_pred_r2) denoting the number of times the R^2 of the full model was smaller than the R^2 of the reduced model across the number of cross-validation iterations, for the predicted test values.

Author

Panagiotis Moulos

Examples

    data(toy,package="prisma")
    data(prisma_out,package="prisma")
    
    response <- "BMI"
    covariates <- c("Case_Control","Gender","Age")
    
    cvm <- prismaCrossValidate(prismaOut,toy,response=response,
        covariates=covariates,leaveOut=c(0.2,0.4),times=5)
#> 
#> ################################################################
#> Running cross-validations for glm with 2 leave-out cases across 1 PRS candidates
#> ################################################################
#> 
#> ================================================================
#> -----> Cross-validating PRS candidate with 24 markers
#> ================================================================
#> 
#> ----- Leave-out percentage 20% -----
#> 
#> Regressing with GLM
#> Trait(s)    : BMI
#> Covariate(s): Case_Control, Gender, Age
#> Regression  : gaussian (predicted)
#> Use PCs     : No
#> 
#> ================================================================
#> -----> Cross-validation iteration 1
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 2
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 3
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 4
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 5
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ----- Leave-out percentage 40% -----
#> 
#> Regressing with GLM
#> Trait(s)    : BMI
#> Covariate(s): Case_Control, Gender, Age
#> Regression  : gaussian (predicted)
#> Use PCs     : No
#> 
#> ================================================================
#> -----> Cross-validation iteration 1
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 2
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 3
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 4
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 5
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ################################################################
#> Running cross-validations for statgen with 2 leave-out cases across 1 PRS candidates
#> ################################################################
#> 
#> ================================================================
#> -----> Cross-validating PRS candidate with 4 markers
#> ================================================================
#> 
#> ----- Leave-out percentage 20% -----
#> 
#> Regressing with GLM
#> Trait(s)    : BMI
#> Covariate(s): Case_Control, Gender, Age
#> Regression  : gaussian (predicted)
#> Use PCs     : No
#> 
#> ================================================================
#> -----> Cross-validation iteration 1
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 2
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 3
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 4
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 5
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ----- Leave-out percentage 40% -----
#> 
#> Regressing with GLM
#> Trait(s)    : BMI
#> Covariate(s): Case_Control, Gender, Age
#> Regression  : gaussian (predicted)
#> Use PCs     : No
#> 
#> ================================================================
#> -----> Cross-validation iteration 1
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 2
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 3
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 4
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 5
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ################################################################
#> Done! Use summarizeCvMetrics() and plotCvMetrics() to summarize and inspect the results.
#> ################################################################
    s <- summarizeCvMetrics(cvm[["glm"]])
#> Summarizing cross-validation statistics for PRS with 24 SNPs
    identical(names(cvm[["glm"]]),names(s)) # TRUE
#> [1] TRUE
    identical(nrow(s[[1]]),2L) # TRUE
#> [1] TRUE
    # Has two leave-out sample percentages 20%, 40%
    
    # Let's display summarized metrics for the 24-SNP PRS 
    # derived by GLM when leaving out 20% of the samples
    # of the initial dataset
    M <- cvm[["glm"]][["24"]][["0.2"]]
    cvSummary(M,nsnp=24,scr=TRUE)
#> 
#> ================================================================
#> PRISMA PRS cross-validation report
#> ================================================================
#> Number of SNPs in PRS                                 : 24
#> Number of cross-validations                           : 5
#> Statistical significance of PRS in model              : 2.185e-09
#> ----------------------------------------------------------------
#> R^2 summary statistics
#> R^2 of the full regression model including the PRS    : 0.7002 +/- 0.0416
#> R^2 of the reduced regression model excluding the PRS : 0.4469 +/- 0.039
#> Adjusted R^2 of the PRS full model contribution       : 0.2532 +/- 0.0499
#> ----------------------------------------------------------------
#> Statistical significance of full R^2 against reduced R^2
#> t-test         : 4.617e-06
#> Wilcoxon test  : 3.968e-03
#> Empirical test : 0
#> ----------------------------------------------------------------
#> RMSE summary statistics
#> RMSE of the full regression model including the PRS    : 2.578 +/- 0.791
#> RMSE of the reduced regression model excluding the PRS : 3.603 +/- 0.7436
#> ----------------------------------------------------------------
#> Statistical significance of full RMSE against reduced RMSE
#> t-test         : 9.661e-01
#> Wilcoxon test  : 9.841e-01
#> Empirical test : 1
#> ----------------------------------------------------------------
#> MAE summary statistics
#> MAE of the full regression model including the PRS    : 1.9377 +/- 0.5275
#> MAE of the reduced regression model excluding the PRS : 2.7673 +/- 0.5733
#> ----------------------------------------------------------------
#> Statistical significance of full MAE against reduced MAE
#> t-test         : 9.777e-01
#> Wilcoxon test  : 9.921e-01
#> Empirical test : 1
#> ----------------------------------------------------------------
#> Correlation summary statistics
#> R between observed & predicted values with the full model    : 0.7943 +/- 0.1286
#> R between observed & predicted values with the reduced model : 0.6044 +/- 0.1684
#> ----------------------------------------------------------------
#> Statistical significance of full R against reduced R
#> t-test         : 4.124e-02
#> Wilcoxon test  : 7.540e-02
#> Empirical test : 0
#> ----------------------------------------------------------------
#> R^2 summary statistics for the predicted test values
#> R^2 from the full regression model including the PRS    : 0.6441 +/- 0.2069
#> R^2 from the reduced regression model excluding the PRS : 0.388 +/- 0.1825
#> Adjusted R^2 from the PRS full model contribution       : 0.2561 +/- 0.148
#> ----------------------------------------------------------------
#> Significance of full R^2 against reduced R^2 (predicted)
#> t-test         : 3.603e-02
#> Wilcoxon test  : 7.540e-02
#> Empirical test : 0
#> ----------------------------------------------------------------
    
    # Some metric plots (ggplot2 graphs)
    cvo <- cvm[[1]]
    p1 <- plotCvMetrics(cvo,what="r2",pl="prs")
#> Creating cross-validation SNP plots for r2
#>   for 0.2 leave out samples fraction
#>   for 0.4 leave out samples fraction
    # p1
    p2 <- plotCvMetrics(cvo,what="rmse",pl="pct")
#> Creating cross-validation percentage plots for rmse
    # p2