Functions to query EBI GWAS and PGS Catalogs

These functions can be used to retrieve PRS candidates in the format of data frames (compatible with the general prisma output as well as the one from aggregatePrsMarkers). This can be done from two sources, the EBI PGS Catalog and the EBI GWAS Catalog given one or more identifiers. These functions build upon the gwasrapidd and quincunx packages which leverage the EBI APIs to fetch the required data. An additional feature regarding PGS scores is that while quincunx only fetches metadata, getPGSScores actually fetches and combines (if possible) the PGS scores, with alleles, effects etc. Then, the resulting data frames can be used with a GWASExperiment object and either a PRISMA external validation pipeline with PRScice2 (takes longer) or with vanilla R regression and cross-validation using e.g. prsCrossValidate to check their performance against one or more traits present in the GWASExperiment object.

Usage

getGWASVariants(efoId = NULL, efoTrait = NULL, 
        removeUnknownRisk = TRUE, retries = 5)
    
    getPGSScores(pgsId = NULL, efoId = NULL, pubmedId = NULL,
        base = NULL, retries = 5, validateLoc = FALSE)
    
    prismaLookup(prismaOut)

Arguments

efoId

One or more EFO ids in proper format. See also Details.

efoTrait

One or more EFO traits in free text format, which should be existing. See also Details.

pgsId

One or more PGS ids in proper format. See also Details.

pubmedId

One or more PubMed ids in proper format.

retries

The number of retries in case of API call failures.

removeUnknownRisk

Should variants of unknown risk be removed in a score from GWAS Catalog? Defaults to TRUE.

base

The base path for the PGS Catalog location. It can be a local path, see also Details.

validateLoc

Validate SNP locations in retrieved PGS scores? Defaults to FALSE, see also Details.

prismaOut

An output object from prisma or a list of PRS candidate data frames.

Value

A data frame such as the one from aggregatePrsMarkers when used with a GWASExperiment object as accompanying input, with as many values filled as possible (e.g. GWAS Catalog does not record variant regression effects.) For prismaLookup, a list of lists. The outer level is named with the GWA methods in prismaOut. The inner level is named according to the PRS candidate names for each GWA method in prismaOut.

Details

Regarding the efoId argument, this has to be one or more valid EFO ids from the GWAS Catalog in proper format (e.g. "EFO_0004566"). Improper format is detected and rejected by the respective function. The same applies to the pgsId argument which should be like "PGS000034" and the pubmedId format which should be like "32527150".

Regarding the base argument, this can be either NULL (the default) to retrieve the PGS score from the link present in the quincunx API call metadata, or a local path following the same structure as the PGS catalog FTP storage. The first case is more guaranteed but prone to connection errors, depending also on the operation system, and especially when multiple score files must be retrieved and processed. The second case assumes that the scores of the PGS Catalog are stored somewhere locally (they do not occupy much space). This can be done for example with Linux wget command:

        mkdir -p LOCAL_PATH
        cd LOCAL_PATH
        wget -m ftp://anonymous:@ftp.ebi.ac.uk/pub/databases/spot/pgs/scores/
        
        # Then in R:
        # base <- LOCAL_PATH/ftp.ebi.ac.uk/pub/databases/spot/pgs/scores/
        # If the long path is not desired:
        
        cd LOCAL_PATH/ftp.ebi.ac.uk/pub/databases/spot/pgs/scores/
        mv * ../../../../../../
        cd LOCAL_PATH
        rm -r ./ftp.ebi.ac.uk/pub/databases/spot/pgs/scores/
        
        # Then in R: base <- LOCAL_PATH
    

Regarding the validateLoc argument, in case of duplicates remaining after basic position/id/risk allele deduplication, when TRUE, performs location validation, as in some cases, duplicate rs ids have different locations (especially when the genome assembly is not reported ("nr") and potentially indicating very old genome assemblies (<hg18). The process relies on the package biomaRt and uses it to query the duplicate rs ids and cross-validate the locations retrieved from Ensembl variation databases. SNPs with locations that cannot be validated are dropped and deduplicated. As the process involves calling Ensembl APIs, it ma be slow. Therefore, if the intended use of the retrieved PRS is to plugin the SNPs to other GWAS data, validateLoc can be FALSE to speed up the process and reduce failures, as the unique SNP ids are used with local genotypes instead of location. In other cases where strict validation is required (e.g. when ORs must be used or reported, then it should be TRUE.

Author

Panagiotis Moulos

Examples

# \donttest{
    # Assume an internet connection!
    efoId <- c("EFO_0004566","EFO_0005937")
    efoTrait <- c("body weight gain",
        "longitudinal BMI measurement") 
    
    test1 <- getGWASVariants(efoId=efoId,
        removeUnknownRisk=FALSE)
#> Scheduling 2 GWAS API calls...
#>   Requesting 2 GWAS API calls with  EFO ID(s)...
#>     Call 1: EFO_0004566
#>     Call 2: EFO_0005937
    test2 <- getGWASVariants(efoTrait=efoTrait)
#> Scheduling 2 GWAS API calls...
#>   Requesting 2 GWAS API calls with EFO trait(s)...
#>     Call 1: body weight gain
#>     Call 2: longitudinal BMI measurement
    
    pgsId <- "PGS000034"
    test3 <- getPGSScores(pgsId=pgsId)
#> Scheduling 1 PGS API calls...
#>   Requesting 1 PGS Catalog API calls with PGS ID(s)...
#>     Call 1: PGS000034
#> Retrieving and enriching scores file(s) for PGS000034
#>   remote file PGS000034.txt.gz
#>   processing PGS000034
#> Dealing with duplicates (if any)
    
    data(prisma_out,package="prisma")
    lookup <- prismaLookup(prismaOut)
#> -----> Querying for glm
#> Querying GWAS Catalog with the largest PRS
#>   Querying association 62070264
#> -----> Querying for statgen
#> Querying GWAS Catalog with the largest PRS
# }