Aggregate PRS candidate SNPs from iterative extraction

This function implements various genotype and sample filters based on properties of the genotypic data present in the input GWAS object. Most filters are losely based in the protocol by Reed et al., 2015 (PubMed ID: 26343929).

Usage

filterGWAS(obj, filters = getDefaults("filters"),
        imputeMissing = TRUE, imputeMode = c("single","split"),
        withPlink = FALSE, rc = NULL)

Arguments

obj

An object of class GWASExperiment holding the genotypic data to be filtered.

filters

A set of genotype filters, see also Details and getDefaults.

imputeMissing

Impute the remaining missing values after filtering? An iterative imputation method based on the package snpStats is used, see also imputeGWAS.

imputeMode

Perform the imputation within the same dataset ("single", default) or split the dataset into a part containing missing values and a part with no missing values ("split"), may happen in large dataset and could be more accurate.

withPlink

Perform basic filtering (not IBD or PCA) with PLINK instead of snpStats. Maybe faster and suitable for huge datasets but with less options. Defaults to FALSE. Also, in this case, the filterRecord is not filled.

rc

The fraction of available cores to use for parallel calculations. Default is NULL for no parallelization.

Value

A filtered GWASExperiment object with or without missing values according to imputeMissing.

Details

Regarding the filters argument, this is a list whose members denote SNP or sample-based filters and their values are filter cutoffs. A more detailed explanation of these filters can be found in Reed et al., 2015 (PubMed ID: 26343929).

The following filters are available:

snpCallRate: The SNPs call rate within a sample. Defaults to 0.98.

sampleCallRate: The sample call rate of SNPs accross samples. Defaults to 0.95.

maf: The Minimum Allele Frequency across samples, below which SNPs are filtered. Defaults to 0.05.

hwe: The Hardy-Weinberg equilibrium p-value. Defaults to 1e-6.

heteroStat: The statistic with which heterozygosity distribution center is defined. Defaults to "median".

heteroFac: The factor with which the scale (standard deviation, MAD) of the heterozygosity distribution is multiplied and added before and after the heteroStat to create a range of acceptable heterozygosity values, e.g. mean +/- 3*std. Defaults to 3. If heteroStat="mean", the scale is the standard deviation. If heteroStat="median", the scale is the MAD.

heteroHard: A hard filter for heterozygosity values. If set, heteroStat and heteroFac are ignored.

pcaOut: Automatically remove samples based on PCA filtering? Defaults to TRUE but should be avoided for datasets with a few samples. Also, it is applied if a robust PCA method is chosen.

pcaRobust: The robust PCA method to be applied for sample exclustion. One of "grid" or "hubert" (default). See the package rrcov for more details.

nPC: Fixed number of PCs to use for PCA filtering. If NA (default), they are determined using the twTest.

LD: Linkage Disequilibrium filter (calculated using package SNPRelate fuctions). Defaults to 0.2.

IBD: Identity By Descent filter (calculated using package SNPRelate fuctions). Defaults to 0.2.

inbreed: Inbreed coefficient filter. Defaults to 0.1.

See also the output of getDefaults("filters").

Author

Panagiotis Moulos

Examples

    data(toy,package="prisma")
    gdsfile(toy) <- system.file(package="prisma",
        "extdata","toy.gds")

    # Lose filters for the toy dataset
    filts <- getDefaults("filters")
    filts$IBD <- NA
    filts$hwe <- 1e-3
    filts$pcaOut <- FALSE
    filts$inbreed <- NA
    
    toyF <- filterGWAS(toy,filters=filts,
        imputeMissing=FALSE)
#> 
#> Performing basic filtering
#>   Calculating SNP summaries for filtering
#>   SNP filters
#>     Call rate
#>     Minor Allele Frequency
#>     Hardy-Weinberg equilibrium
#>   Sample filters
#>     Call rate
#>     Heterozygosity
#>   Summarizing SNP filter results
#>   Summarizing sample filter results
#> 
#> ----- Filtering report: -----
#> Filtered SNPs:
#>   SNP call rate             : 0
#>   Minor Allele Frequency    : 119
#>   Hardy-Weinberg equilibrium: 2
#> 
#> Filtered samples:
#>   Sample call rate          : 0
#>   Heterozygosity            : 0
#>   Inbreed                   : 0
#> 
        
    # With PLINK
    if (Sys.which("plink") != "")
        toyF <- filterGWAS(toy,filters=filts,
            imputeMissing=FALSE,withPlink=TRUE)
#> 
#> Performing basic filtering
#>   Exporting temporary PLINK files for the calculations
#> Preflight...
#> Writing FAM file to /tmp/Rtmps2YAD0/file14c2c93352d.fam 
#> Writing extended MAP file to /tmp/Rtmps2YAD0/file14c2c93352d.bim 
#> Writing BED file to /tmp/Rtmps2YAD0/file14c2c93352d.bed (SNP-major mode)
#> 
#> Executing:
#> /home/panos/.local/share/R/plink-1.90/plink \
#>   --bfile /tmp/Rtmps2YAD0/file14c2c93352d \
#>   --out /tmp/Rtmps2YAD0/file14c7b010438 --het
#> 
#> Executing:
#> /home/panos/.local/share/R/plink-1.90/plink \
#>   --bfile /tmp/Rtmps2YAD0/file14c2c93352d \
#>   --out /tmp/Rtmps2YAD0/file14c7b010438 \
#>   --make-bed \
#>   --geno 0.02 \
#>   --maf 0.05 \
#>   --hwe 0.001 \
#>   --mind 0.05
#> Reading PLINK files with snpStats framework
#> 881 SNPs and 100 samples remaining