Functions to perform GWA tests with various algorithms

These functions perform the association tests to create summary statistics for a GWA dataset prior to the extraction of PRS candidates. They are part of the overall PRISMA workflow (prisma) but they are designed to be also used independently just to look for univariate association cases, i.e. look for associations for a single trait using various tests.

Usage

gwa(obj, response, covariates = NULL, pcs = FALSE,
        psig = 0.05, methods = c("glm", "rrblup", "statgen",
            "snptest", "plink", "lasso"),
        combine = c("fisher", "simes", "max", "min", "harmonic",
            "whitlock","pandora"), glmOpts = getDefaults("glm"),
        rrblupOpts = getDefaults("rrblup"), 
        statgenOpts = getDefaults("statgen"),
        snptestOpts = getDefaults("snptest"),
        plinkOpts = getDefaults("plink"), rc = NULL, ...)
    
    gwaBlup(obj, response, covariates = NULL, 
        usepc = c("auto", "estim", "rrint", "fixed", "none"), 
        npcs = NULL, psig = 0.05, rc = NULL)
    
    gwaGlm(obj, response, covariates = NULL, pcs = FALSE,
        family = NULL, psig = 0.05, penalized = FALSE,
        size = 1, rc = NULL, ...)
        
    gwaPlink(obj, response, covariates = NULL, pcs = TRUE,
        psig = 0.05, seed = 42, cliopts = NULL, wspace = NULL, 
        rc = NULL)
    
    gwaStatgen(obj, response, covariates = NULL, pcs = TRUE, 
        psig = 0.05, rc = NULL)
    
    gwaSnptest(obj, response, covariates = NULL, pcs = TRUE,
        psig = 0.05, test = c("frequentist", "bayesian"), 
        model = c("additive", "dominant", "recessive", 
        "general", "heterozygote"), wspace = NULL)

Arguments

obj

an object of class GWASExperiment holding the genotypic and phenotypic data to be used for association tests.

response

The trait for which associations are searched for. It should be a character of length 1 and exist in the phenotypes of obj.

covariates

The covariates to include in association methods/regression models. It should be a character vector of traits all existing in the phenotypes of obj.

pcs

Inlcude PCs in the association models? Defaults to FALSE.

psig

The significance level (p-value cutoff) for which an association between a SNP/marker and a phenotype (response) is considered significant. Defaults to 0.05.

methods

The GWA methods to apply on obj with respect to response and covariates. It can be one or more of "glm", "rrblup", "statgen", "snptest", "plink", "lasso". See also Details for explanation of each case.

combine

How (if) p-values should be combined to produce a summarized p-value from all GWA methods for each association test. Not all methods are yet available, see also Details.

glmOpts

A list of options for GLM regression model fit. See also Details.

rrblupOpts

A list of options for ridge regression model fit (rrBLUP). See also Details.

statgenOpts

A list of options for regression fit with statgenGWAS. See also Details.

snptestOpts

A list of options for modeling associations with SNPTEST. See also Details.

plinkOpts

A list of options for modeling associations with PLINK. See also Details.

usepc

What to do with PCA and PCs when using gwaBlup and rrBLUP model fitting. Defaults to "auto" and also "rrint" is preferred. See also Details.

npcs

Number of PCs to use with rrBLUP. Defaults to NULL for auto-determining what to do.

family

The model family for GLM, e.g. "gaussian" for classic regression with continuous response variable or "binomial" for logistic regression. By default (NULL), it is guessed from the form of the response variable.

penalized

Use penalized (LASSO) regression for GLM. Defaults to FALSE.

size

The number of SNPs to include in a GLM model, usually utilized with penalized=TRUE. Defaults to 1 for models with 1 SNP as covariate.

wspace

A path to create workspaces and temporary files for SNPTEST and PLINK which operate outside R.

test

One of supported test types for SNPTEST. It can be "frequentist" (default) or "bayesian". See also SNPTEST documentation.

seed

A random seed for PLINK resampling operations. Defaults to 42.

cliopts

A string with additional command-line parameters to pass to the PLINK executable. Defaults to NULL.

model

One of the supported genotype modeling types for SNPTEST. It can be one of "additive" (default), "dominant", "recessive", "general" or "heterozygote". See also SNPTEST documentation.

rc

The fraction of available cores to use for parallel calculations. Default is NULL for no parallelization.

...

Additional arguments passed to regression functions.

Value

Depending on the function used, either a data frame with summary statistics (always with at least SNP name, effects/ORs and p-values)

Details

Regarding the methods argument, the following briefly explain how the associations are modeled in each case:

glm: In this case, a SNP is checked for association with the response trait using a General Linear Regression model of the form: response ~ covariates + [SNP] or response ~ covariates + pcs + [SNP] depending on whether PCs are added as covariates or not. [SNP] is simply the recorded genotype (0 for homozygous minor allele, 1 for heterozygous allele and 2 for homozygous major allele). The p-value and the regression coefficient (effect/OR) of the [SNP] is recorded and used to determine if the SNP is associated with the covariate or not. The function can be used independently by directly using gwaGlm. In this case, a data frame of summary statistics is returned instead of the input GWASExperiment.

rrblup: In this case, the GWAS function from the package rrBLUP is used to model the associations, the effects and the significance of each SNP using Ridge Regression. The function can be used independently by directly using gwaBlup. In this case, a data frame of ridge regression summary statistics is returned instead of the input GWASExperiment.

statgen: In this case, the runSingleTraitGwas function from the statgenGWAS package is used to model the associations, the effects and the significance of each SNP using also linear regression. The function can be used independently by directly using gwaStatgen. In this case, a data frame of summary statistics is returned instead of the input GWASExperiment.

snptest: In this case, the SNPTEST software is used to check for associations and summary statistics. It should be noted that SNPTEST operates with at least 100 samples. The function can be used directly as the previous ones by gwaSnptest.

plink: In this case, the PLINK software is used to check for associations and return summary statistics. The function can be used directly as the previous ones by gwaPlink.

Regarding the combine argument, this is used to control how the p-values returned by the usage of many association tests will be combined to produce a summarized p-value. Currently "whitlock", "pandora" and "harmonic" are placeholders since the estimation of p-value weights are work in progress. For the rest, please see Moulos and Hatzis, 2015 (PMID: 25452340) and Fanidis and Moulos, 2021 (PMID: 32778872).

Regarding the glmOpts argument, this is a named list with members:

family: The regression family, which defaults to NULL for auto-detection.

size: The number of sequential SNPs to include in a lasso regression model (currently in testing mode).

Regarding the rrblupOpts argument, this is a named list with members:

pcblup: How should the PCs included in ridge regression be calculated? Defaults to "auto", see the main function arguments for the rest of the options.

npcs: The number of PCs to include.

Regarding the statgenOpts argument, this is a named list with members:

kinship: The kinship matrix estimation method, to be included in the statgenGWAS model fitting. Defaults to "astle", see the statgenGWAS package for other options.

reml: The method for calculated the Restricted Maximum Likelihood in the statgenGWAS model fitting. Defaults to "EMMA", see the statgenGWAS package for other options.

gls: The method used to estimate the SNP effects, "signle" (default) for using a single kinship matrix, or "multi" for using chromosome specific kinship matrices. See the statgenGWAS package for further explanations.

Regarding the snptestOpts argument, this is a named list with members:

test: The type of test to be used (see also main arguments above). Defaults to "frequentist". See also SNPTEST documentation.

model: The genotype effect estimation mode, defaults to "additive", See also main arguments and SNPTEST documentation.

workspace: The path to SNPTEST temporary workspace. Defaults to NULL for using tempdir. It should be provided for maximum control.

Regarding the plinkOpts argument, this is a named list with members:

effect: How should the effects be modeled for each SNP (similar to the model of SNPTEST). Please see PLINK documentation for a full list of options.

seed: A seed for PLINK randomized operations such as resampling and permuting. Defaults to 42.

cliopts: Further options to be passed to plink command, as these would be given to the command line. Please note that these should be switched and options accepting numbers or srings but _not_ input/output files.

Regarding the usepc argument of gwaBlup, this can be one of "auto", "estim", "rrint", "fixed" or "none". It controls how rrBLUP models use the PCs in the generic ridge regression model. The following are available:

"auto": When in automated mode, if the input object contains calculated PCs to be used as covariates, these are used, otherwise the optimal number is estimate (switch to "estim".)

"estim": Automatic iterative estimation.

"rrint": In this case, the number of PCs to include are taken from existing PCs in the input GWASExperiment object, or from the npcs argument if there are no PCs in the input. Then, this number is used internally to the GWAS rrBLUP function to include PCs in the model. Although not the default, it seems that this is the more stable method and it is recommended to be used.

"fixed": This uses npcs only (in a similar way as "auto".)

"none": Do not include PCs in the model.

Author

Panagiotis Moulos

Examples

    data(toy,package="prisma")
    gdsfile(toy) <- system.file(package="prisma","extdata","toy.gds")
    
    # Only 1 PC will be returned because of dataset size
    toy <- suppressWarnings(calcPcaCovar(toy,method="snprel"))
#> Performing PCA to select covariates to include in associations
#> Reading GDS file /tmp/RtmpezHjqb/temp_libpath372f48f1c2bc/prisma/extdata/toy.gds...
#> Performing LD pruning...
#> SNP pruning based on LD:
#> Excluding 0 SNP (non-autosomes or non-selection)
#> Excluding 38 SNPs (monomorphic: TRUE, MAF: NaN, missing rate: NaN)
#>     # of samples: 100
#>     # of SNPs: 962
#>     using 1 thread
#>     sliding window: 500,000 basepairs, Inf SNPs
#>     |LD| threshold: 0.2
#>     method: composite
#> Chromosome 18: 23.44%, 154/657
#> Chromosome 21: 21.28%, 73/343
#> 227 markers are selected in total.
#> LD pruning finished! 227 SNPs will be used for PCA.
#> 
#> Performing PCA...
#> Principal Component Analysis (PCA) on genotypes:
#> Excluding 773 SNPs (non-autosomes or non-selection)
#> Excluding 0 SNP (monomorphic: TRUE, MAF: NaN, missing rate: NaN)
#>     # of samples: 100
#>     # of SNPs: 227
#>     using 1 thread
#>     # of principal components: 32
#> PCA:    the sum of all selected genotypes (0,1,2) = 10268
#> CPU capabilities: Double-Precision SSE2
#> Mon Jun  6 20:06:52 2022    (internal increment: 13024)
#> [..................................................]  0%, ETC: ---        [==================================================] 100%, completed, 0s
#> Mon Jun  6 20:06:52 2022    Begin (eigenvalues and eigenvectors)
#> Mon Jun  6 20:06:52 2022    Done.
#> Selecting number of PCs to include
#>   with Tracy-Widom test for significant eigenvalues
#>     selected 1 PCs
    
    toyG <- gwa(toy,response="BMI",covariates=c("Age","Gender"),
        pcs=TRUE,psig=0.05,methods=c("glm","statgen"))
#> 
#> Starting GWA analysis in 2 chunks with 2 methods: glm, statgen
#> 
#> ========== Analyzing chromosome/part 18
#> 
#> Performing GWAS with GLM over 1 chunks
#> Trait(s)    : BMI
#> Covariate(s): Age, Gender
#> Regression  : gaussian (predicted)
#> Use PCs     : Yes
#>   testing SNPs from rs1009238 to rs1568327
#> Found 24 associations (uncorrected p-values).
#> 
#> Preparing data for statgenGWAS...
#> Performing GWAS with statgenGWAS model
#> Trait(s)    : BMI
#> Covariate(s): Age, Gender
#> Use PCs     : Yes
#> Found 24 associations (uncorrected p-values).
#> 
#> Combining p-values from multi-GWAS using fisher method
#> 
#> Overall, found 54 associations with (uncorrected) combined p-values).
#> 
#> ========== Analyzing chromosome/part 21
#> 
#> Performing GWAS with GLM over 1 chunks
#> Trait(s)    : BMI
#> Covariate(s): Age, Gender
#> Regression  : gaussian (predicted)
#> Use PCs     : Yes
#>   testing SNPs from rs2824974 to rs2828316
#> Found 18 associations (uncorrected p-values).
#> 
#> Preparing data for statgenGWAS...
#> Performing GWAS with statgenGWAS model
#> Trait(s)    : BMI
#> Covariate(s): Age, Gender
#> Use PCs     : Yes
#> Found 18 associations (uncorrected p-values).
#> 
#> Combining p-values from multi-GWAS using fisher method
#> 
#> Overall, found 35 associations with (uncorrected) combined p-values).
#> 
#> Finished, the following putative associations were found per part:
#>   Part 18: 54 associations found out of 657 markers
#>   Part 21: 35 associations found out of 343 markers
#> 
#> Overall, found 89 associations with (uncorrected) combined p-values).
    
    glmSumstat <- gwaGlm(toy,response="BMI",
        covariates=c("Age","Gender"),pcs=TRUE)
#> 
#> Performing GWAS with GLM over 1 chunks
#> Trait(s)    : BMI
#> Covariate(s): Age, Gender
#> Regression  : gaussian (predicted)
#> Use PCs     : Yes
#>   testing SNPs from rs1009238 to rs2828316
#> Found 42 associations (uncorrected p-values).
    
    rrSumstat <- gwaBlup(toy,response="BMI",
        covariates=c("Age","Gender"),usepc="rrint",npcs=1)
#> 
#> Preparing genotypes for rrBLUP...
#> Performing GWAS with rrBLUP model
#> Trait(s)    : BMI
#> Covariate(s): Age, Gender
#> Use PCs     : Yes
#> Found 55 associations (uncorrected p-values).
    
    sgSumstat <- gwaStatgen(toy,response="BMI",
        covariates=c("Age","Gender"),pcs=TRUE)
#> 
#> Preparing data for statgenGWAS...
#> Performing GWAS with statgenGWAS model
#> Trait(s)    : BMI
#> Covariate(s): Age, Gender
#> Use PCs     : Yes
#> Found 42 associations (uncorrected p-values).
    
    if (Sys.which("snptest") != "")
        snpSumstat <- gwaSnptest(toy,response="BMI",
            covariates=c("Age","Gender"),pcs=TRUE)
#> snptest operations will be executed at /tmp/Rtmpxo9Dha/snptest_AZQIL3254L
#> Preparing SNPTEST association run at workspace directory: /tmp/Rtmpxo9Dha/snptest_AZQIL3254L
#>  writing SNPTEST sample file in /tmp/Rtmpxo9Dha/snptest_AZQIL3254L
#>   writing PLINK files in /tmp/Rtmpxo9Dha/snptest_AZQIL3254L
#> Preflight...
#> Writing FAM file to /tmp/Rtmpxo9Dha/snptest_AZQIL3254L/plink_snptest_DPKHE3458Z.fam 
#> Writing extended MAP file to /tmp/Rtmpxo9Dha/snptest_AZQIL3254L/plink_snptest_DPKHE3458Z.bim 
#> Writing BED file to /tmp/Rtmpxo9Dha/snptest_AZQIL3254L/plink_snptest_DPKHE3458Z.bed (SNP-major mode)
#> Preparation done!
#> 
#> Performing GWAS with SNPTEST
#> Trait(s)    : BMI
#> Covariate(s): Age, Gender
#> Use PCs     : Yes
#> 
#> Executing:
#> /home/panos/.local/share/R/snptest-2.5.4/snptest \
#>   -data /tmp/Rtmpxo9Dha/snptest_AZQIL3254L/plink_snptest_DPKHE3458Z.bed \
#>     /tmp/Rtmpxo9Dha/snptest_AZQIL3254L/pheno.sample \
#>   -frequentist 1 \
#>   -method score \
#>   -pheno BMI \
#>   -cov_all \
#>   -o /tmp/Rtmpxo9Dha/snptest_AZQIL3254L/snptest.out
#> Found 38 associations (uncorrected p-values).
    
    if (Sys.which("plink") != "")
        plSumstat <- gwaPlink(toy,response="BMI",
            covariates=c("Age","Gender"),pcs=TRUE)
#> plink operations will be executed at /tmp/Rtmpxo9Dha/plink_PIICY8094S
#> Preparing PLINK association run at workspace directory: /tmp/Rtmpxo9Dha/plink_PIICY8094S
#>   writing the PLINK phenotype file...
#>   writing the covariates file...
#>   writing PLINK genotype files in /tmp/Rtmpxo9Dha/plink_PIICY8094S
#> Preflight...
#> Writing FAM file to /tmp/Rtmpxo9Dha/plink_PIICY8094S/plink_plink_QJCXD6659L.fam 
#> Writing extended MAP file to /tmp/Rtmpxo9Dha/plink_PIICY8094S/plink_plink_QJCXD6659L.bim 
#> Writing BED file to /tmp/Rtmpxo9Dha/plink_PIICY8094S/plink_plink_QJCXD6659L.bed (SNP-major mode)
#> Preparation done!
#> 
#> Performing GWAS with PLINK
#> Trait(s)    : BMI
#> Covariate(s): Age, Gender
#> Use PCs     : Yes
#> 
#> Executing:
#> /home/panos/.local/share/R/plink-1.90/plink \
#>   --bfile /tmp/Rtmpxo9Dha/plink_PIICY8094S/plink_plink_QJCXD6659L \
#>   --pheno /tmp/Rtmpxo9Dha/plink_PIICY8094S/pheno_QJCXD6659L \
#>   --allow-no-sex \
#>   --ci 0.95 \
#>   --seed 42 \
#>   --out /tmp/Rtmpxo9Dha/plink_PIICY8094S/out_QJCXD6659L \
#>   --covar /tmp/Rtmpxo9Dha/plink_PIICY8094S/covar_QJCXD6659L \
#>   --linear perm
#> Found 43 associations (uncorrected p-values).