Functions to convert GWASExperiment to other classes

These functions either convert a GWASExperiment to another object or file required for functionalities in the supporting packages, or convert between human genome coordinate systems. Finally, to facilitate this conversion process when required, a functionality to "guess" the human genome version is provided.

Usage

GWASExperiment2gData(obj, covariates = NULL, pcs = FALSE,
        reverse = FALSE)
    
    GWASExperiment2GDS(obj)
    
    GWASExperimentLiftOver(obj, from, to)
    
    guessHumanGenomeVersion(obj)

Arguments

obj

a GWASExperiment object.

covariates

the covariates to include in the gData object (package statgenGWAS) including the response phenotype.

pcs

include the PCs present in obj in the gData object? Defaults to FALSE.

reverse

reverse the genotypes in the construction of a gData object? Defaults to FALSE.

from

the source genome version, one of "hg19" or "hg38".

to

the target genome version, one of "hg19" or "hg38".

Value

For GWASExperiment2gData, a gData object that can be used with functions from the statgenGWAS package. For GWASExperiment2GDS, a GDS file is written to the location specified by gdsfile(obj). For GWASExperimentLiftOver, the same GWASExperiment object is returned, with changed SNP positions from one genome version to the other. Some SNPs may be dropped because of inability of the liftOver algorithm to map between versions (rare). For guessHumanGenomeVersion, a string denoting the guessed human genome version ("hg19" or "hg38").

Details

Regarding the guessHumanGenomeVersion function, it queries dbSNP for 5 random SNPs present in obj using the package rsnps. Depending on the result, the genome version is guessed.

Author

Panagiotis Moulos

Examples

    data(toy,package="prisma")
    
    # There are some non-numeric phenotypes - ignore
    gd <- suppressWarnings(GWASExperiment2gData(toy))
    
    gdsfile(toy) <- tempfile()
    GWASExperiment2GDS(toy)
#> Preflight...
#> Writing FAM file to /tmp/Rtmps2YAD0/file14c681a9e70.fam 
#> Writing extended MAP file to /tmp/Rtmps2YAD0/file14c681a9e70.bim 
#> Writing BED file to /tmp/Rtmps2YAD0/file14c681a9e70.bed (SNP-major mode)
#> Start file conversion from PLINK BED to SNP GDS ...
#>     BED file: '/tmp/Rtmps2YAD0/file14c681a9e70.bed'
#>         SNP-major mode (Sample X SNP), 24.4K
#>     FAM file: '/tmp/Rtmps2YAD0/file14c681a9e70.fam'
#>     BIM file: '/tmp/Rtmps2YAD0/file14c681a9e70.bim'
#> Wed Jun 29 19:38:12 2022     (store sample id, snp id, position, and chromosome)
#>     start writing: 100 samples, 1000 SNPs ...
#> [..................................................]  0%, ETC: ---        [==================================================] 100%, completed, 0s
#> Wed Jun 29 19:38:12 2022 	Done.
#> Optimize the access efficiency ...
#> Clean up the fragments of GDS file:
#>     open the file '/tmp/Rtmps2YAD0/file14c2393c7e1' (34.6K)
#>     # of fragments: 50
#>     save to '/tmp/Rtmps2YAD0/file14c2393c7e1.tmp'
#>     rename '/tmp/Rtmps2YAD0/file14c2393c7e1.tmp' (34.3K, reduced: 312B)
#>     # of fragments: 24
    file.exists(gdsfile(toy)) # TRUE
#> [1] TRUE
    
    genome(toy) # hg19
#> [1] "hg19"
    toy2 <- GWASExperimentLiftOver(toy,from="hg19",
        to="hg38")
#> Lifting over features
#>   lifting chr18
#>   lifting chr21
    
    guessHumanGenomeVersion(toy) # hg19
#> Getting info about the following rsIDs: rs238146, rs273746, rs13048320, rs4798085
#> [1] "hg19"