Create GWASExperiments from PLINK files

This function creates a GWASExperiment object from PLINK BED/BIM/FAM files using functions from the package snpStats. In the future, importing with the package bigSNP will be supported.

Usage

importGWAS(input, phenos = NULL, backend = c("snpStats",
        "bigsnpr"), selection = NULL, genome = NA_character_,
        alleleOrder = c("plink", "reverse"), writeGds = TRUE,
        gdsfile = ifelse(backend == "snpStats", tempfile(), NA),
        gdsOverwrite = TRUE)

Arguments

input

A list with the full path to a PLINK file triplet (BED/BIM/FAM) or a string with the triplet basename which will be suffixed with bed/bim/fam. See als Details.

phenos

A data frame or a file with the dataset phenotypes. It must have the same number of rows as the samples present in the PLINK triplet.

backend

the R backend to use for object management. See Details.

selection

A list with members named snps and samples to be used as the related select.snps and select.subjects arguments of the read.plink function from snpStats.

genome

The reference genome which the SNPs are mapped to. Typically one of "hg19" or "hg38" if this information is available, otherwise NA.

alleleOrder

Should the alleles be reversed? If yes choose "reverse".

writeGds

Write a GDS supporting file while importing PLINK files? In some cases such as reading files for external imputation is not needed. Defaults to TRUE.

gdsfile

A GDS file supported by the package SNPRelate is necessary for some operations like SNPRelate PCA and LD. Where should this be written? It is adivsed to explicitly provide this path.

gdsOverwrite

overwrite the provided GDS file if exists? Defaults to TRUE.

Value

A GWASExperiment object with the contents of the PLINK files triplet and optionally, additional associated phenotypes.

Details

Regarding the input argument, when this is a list, the latter must have named members with names bed, bim, fam. When this is a single string, it must point to a full path where the triplet is found and must not end with any of bim, bed, fam as these are appended to the base.

Regarding the backend argument, current PRISMA design goals are to support both importing and processing data with the R package snpStats and the R package bigSNP. Currently, only snpStats is supported.

Author

Panagiotis Moulos

Examples

    fam <- system.file("extdata/sample.fam",package="snpStats")
    bim <- system.file("extdata/sample.bim",package="snpStats")
    bed <- system.file("extdata/sample.bed",package="snpStats")
    input <- list(fam=fam,bim=bim,bed=bed)
    sample <- snpStats::read.plink(input$bed,input$bim,input$fam)
    
    # Without phenotypes
    gwe <- importGWAS(input,backend="snpStats")
#> Reading PLINK files with snpStats framework
#> Reading PLINK files with SNPRelate framework and storing output tp /tmp/Rtmps2YAD0/file14c7c0de2bb
#> Start file conversion from PLINK BED to SNP GDS ...
#>     BED file: '/home/panos/R/x86_64-pc-linux-gnu-library/4.1/snpStats/extdata/sample.bed'
#>         SNP-major mode (Sample X SNP), 603 bytes
#>     FAM file: '/home/panos/R/x86_64-pc-linux-gnu-library/4.1/snpStats/extdata/sample.fam'
#>     BIM file: '/home/panos/R/x86_64-pc-linux-gnu-library/4.1/snpStats/extdata/sample.bim'
#> Wed Jun 29 19:38:24 2022     (store sample id, snp id, position, and chromosome)
#>     start writing: 120 samples, 20 SNPs ...
#> [..................................................]  0%, ETC: ---        [==================================================] 100%, completed, 0s
#> Wed Jun 29 19:38:24 2022 	Done.
#> Optimize the access efficiency ...
#> Clean up the fragments of GDS file:
#>     open the file '/tmp/Rtmps2YAD0/file14c7c0de2bb' (5.1K)
#>     # of fragments: 50
#>     save to '/tmp/Rtmps2YAD0/file14c7c0de2bb.tmp'
#>     rename '/tmp/Rtmps2YAD0/file14c7c0de2bb.tmp' (4.8K, reduced: 312B)
#>     # of fragments: 24
    
    # With phenotypes
    set.seed(42)
    pseudopheno <- data.frame(
        case_control=sample(c(0,1),nrow(sample$fam),replace=TRUE),
        other_pheno=sample(c("drug","nodrug"),nrow(sample$fam),replace=TRUE),
        cont=round(runif(nrow(sample$fam)),3),
        row.names=rownames(sample$fam)
    )
    gwe <- importGWAS(input,phenos=pseudopheno,backend="snpStats")
#> Reading PLINK files with snpStats framework
#> Reading PLINK files with SNPRelate framework and storing output tp /tmp/Rtmps2YAD0/file14c6a724626
#> Start file conversion from PLINK BED to SNP GDS ...
#>     BED file: '/home/panos/R/x86_64-pc-linux-gnu-library/4.1/snpStats/extdata/sample.bed'
#>         SNP-major mode (Sample X SNP), 603 bytes
#>     FAM file: '/home/panos/R/x86_64-pc-linux-gnu-library/4.1/snpStats/extdata/sample.fam'
#>     BIM file: '/home/panos/R/x86_64-pc-linux-gnu-library/4.1/snpStats/extdata/sample.bim'
#> Wed Jun 29 19:38:24 2022     (store sample id, snp id, position, and chromosome)
#>     start writing: 120 samples, 20 SNPs ...
#> [..................................................]  0%, ETC: ---        [==================================================] 100%, completed, 0s
#> Wed Jun 29 19:38:24 2022 	Done.
#> Optimize the access efficiency ...
#> Clean up the fragments of GDS file:
#>     open the file '/tmp/Rtmps2YAD0/file14c6a724626' (5.1K)
#>     # of fragments: 50
#>     save to '/tmp/Rtmps2YAD0/file14c6a724626.tmp'
#>     rename '/tmp/Rtmps2YAD0/file14c6a724626.tmp' (4.8K, reduced: 312B)
#>     # of fragments: 24