Genotype imputation for GWASExperiment objects

This function performs internal imputation in a GWASExperiment object using a combined approach with fucntionalities from the snpStats packages and the scrime package. The internal imputation in PRISMA refers to missing genotype imputation within the same dataset and using existing genotypes _within_ the same dataset to impute missing genotypes. It should not be confused with imputation based on an external reference panel such as the 1000 Genomes Project which can be used to both impute and extend a dataset to include non-genotyped SNPs/markers. This process is also available with PRISMA and is refered as "extension" (see also the extendGWAS function.)

Usage

imputeGWAS(obj, mode = c("single", "split"),
        failed = c("scrime", "none"), rc = NULL)

Arguments

obj

A GWASExperiment object with missing genotype values to be imputed.

mode

Perform the imputation within the same dataset ("single", default) or split the dataset into a part containing missing values and a part with no missing values ("split"). See also Details and Description for further explanation of the internal imputation process.

failed

What to do with genotypes that failed to be imputed with the iterative regression process? Impute with kNN from the scrime package ("scrime", default) or no further imputation ("none").

rc

The fraction of available cores to use for parallel calculations. Default is NULL for no parallelization.

Value

A GWASExpriment object with no (failed="scrime") missing values or fewer (failed="none") missing values in the genotypes.

Details

Regarding the mode argument, this controls how the imputation based on snpStats functions (snp.imputation) will be performed. When mode="single", the imputation rules are created within the total initial dataset without any splitting to a "better" and a "worse" dataset in terms of missing value content. When mode="split", the dataset is split in a part containing samples with missing values and a part with samples without missing values. While this is not realistic for smaller datasets it may happen in large dataset and could be more accurate. If splitting is chosen for small datasets, it's likely that the imputation rules will be created with a very small number of samples, resulting in fewer rules (i.e. fewer SNPs can be imputed) and with poorer accuracy.

The internal imputation process in PRISMA is split in two phases:

Phase 1: The genotypes of the GWASExperiment input is used to create imputation rules (with the respective function of snpStats). These rules are essentially linear regression models which model potential predictions of SNP genotypes according to nearby genotypes and present high R^2 values and therefore can be used to provide missing cases with relatively good accuracy. This process is iteratively deployed in PRISMA, until no more imputation rules can be produced, i.e. the result of each imputation iteration is used to create new rules to impute more SNPs if possible, until no more new rules can be produced.

Phase 2: In this phase, the missing genotypes that could not be imputed with the iterative process of Phase 1, are imputed with the k-nearest neighbor algorithm for genotype imputation from the R package scrime. This step is optional.

Phase 3: Missing genotypes that could not be imputed with the scrime package in Phase 2 are replaced with the global average genotype of the rest of the genotypes. This phase is optional and should be quite rare to be needed.

Author

Panagiotis Moulos

Examples

    data(toy,package="prisma")
    
    # Make some genotypes missing
    set.seed(42)
    toMiss <- lapply(seq_len(ncol(toy)),function(j) {
        n <- sample(20,1)
        miss <- logical(nrow(toy))
        miss[sample(nrow(toy),n)] <- TRUE
        return(miss)
    })
    
    G <- as(genotypes(toy),"numeric") + 1
    mode(G) <- "integer"
    for (j in seq_len(ncol(toy)))
        G[toMiss[[j]],j] <- 0L
    G <- SnpMatrix(G)
#> coercing object of mode  numeric  to SnpMatrix
    genotypes(toy) <- G
    
    toyI <- imputeGWAS(toy,fail="none")
#> 
#> Imputing missing values with snpStats rules
#> 
#> Starting imputation analysis in 2 chunks
#> 
#> ========== Imputing chromosome/part 18
#> -----> Imputation iteration 1
#> Creating imputation rules
#> SNPs tagged by a single SNP: 45
#> SNPs tagged by multiple tag haplotypes (saturated model): 537
#> 97 samples have missing genotypes in 433 SNPs in total.
#> Imputing...
#> Applying imputations...
#> -----> Imputation iteration 2
#> Creating imputation rules
#> SNPs tagged by a single SNP: 44
#> SNPs tagged by multiple tag haplotypes (saturated model): 538
#> 52 samples have missing genotypes in 47 SNPs in total.
#> Imputing...
#> Applying imputations...
#> 
#> ========== Imputing chromosome/part 21
#> -----> Imputation iteration 1
#> Creating imputation rules
#> SNPs tagged by a single SNP: 20
#> SNPs tagged by multiple tag haplotypes (saturated model): 301
#> 92 samples have missing genotypes in 221 SNPs in total.
#> Imputing...
#> Applying imputations...
#> -----> Imputation iteration 2
#> Creating imputation rules
#> SNPs tagged by a single SNP: 20
#> SNPs tagged by multiple tag haplotypes (saturated model): 301
#> 27 samples have missing genotypes in 17 SNPs in total.
#> Imputing...
#> Applying imputations...
#> 
#> Imputation finished, re-merging the output
    before <- length(which(is.na(genotypes(toy))))
    after <- length(which(is.na(genotypes(toyI))))
    before > after # TRUE
#> [1] TRUE