Functions to list analyses in a GWASExperiment

These functions list the associated analyses with a GWASExperiment, i.e. GWA tests for association summary statistics, or PRS weights from PRS extraction analysis. A GWASExperiment may have multiple GWA and PRS analysis performed on it, such as GWA with multiple or no PCs as covariates. Each analysis is attached to the main GWASExperiment to avoid repetition of large genotypic data.

Usage

listGwa(obj)
    listPrs(obj)

Arguments

obj

a GWASExperiment object.

Value

These functions only print information on screen.

Author

Panagiotis Moulos

Examples

    data(toy,package="prisma")
    gdsfile(toy) <- system.file(package="prisma","extdata","toy.gds")
    
    # Only 1 PC will be returned because of dataset size
    toy <- suppressWarnings(calcPcaCovar(toy,method="snprel"))
#> Performing PCA to select covariates to include in associations
#> Reading GDS file /tmp/RtmplLIKfg/temp_libpath4f71185f9e1c/prisma/extdata/toy.gds...
#> Performing LD pruning...
#> SNP pruning based on LD:
#> Excluding 0 SNP (non-autosomes or non-selection)
#> Excluding 38 SNPs (monomorphic: TRUE, MAF: NaN, missing rate: NaN)
#>     # of samples: 100
#>     # of SNPs: 962
#>     using 1 thread
#>     sliding window: 500,000 basepairs, Inf SNPs
#>     |LD| threshold: 0.2
#>     method: composite
#> Chromosome 18: 23.44%, 154/657
#> Chromosome 21: 21.28%, 73/343
#> 227 markers are selected in total.
#> LD pruning finished! 227 SNPs will be used for PCA.
#> 
#> Performing PCA...
#> Principal Component Analysis (PCA) on genotypes:
#> Excluding 773 SNPs (non-autosomes or non-selection)
#> Excluding 0 SNP (monomorphic: TRUE, MAF: NaN, missing rate: NaN)
#>     # of samples: 100
#>     # of SNPs: 227
#>     using 1 thread
#>     # of principal components: 32
#> PCA:    the sum of all selected genotypes (0,1,2) = 10268
#> CPU capabilities: Double-Precision SSE2
#> Fri Jun  3 14:30:10 2022    (internal increment: 13024)
#> [..................................................]  0%, ETC: ---        [==================================================] 100%, completed, 0s
#> Fri Jun  3 14:30:10 2022    Begin (eigenvalues and eigenvectors)
#> Fri Jun  3 14:30:10 2022    Done.
#> Selecting number of PCs to include
#>   with Tracy-Widom test for significant eigenvalues
#>     selected 1 PCs
    
    toy <- gwa(toy,response="BMI",covariates=c("Age","Gender"),
        pcs=TRUE,psig=0.05,methods=c("glm","statgen"))
#> 
#> Starting GWA analysis in 2 chunks with 2 methods: glm, statgen
#> 
#> ========== Analyzing chromosome/part 18
#> 
#> Performing GWAS with GLM over 1 chunks
#> Trait(s)    : BMI
#> Covariate(s): Age, Gender
#> Regression  : gaussian (predicted)
#> Use PCs     : Yes
#>   testing SNPs from rs1009238 to rs1568327
#> Found 24 associations (uncorrected p-values).
#> 
#> Preparing data for statgenGWAS...
#> Performing GWAS with statgenGWAS model
#> Trait(s)    : BMI
#> Covariate(s): Age, Gender
#> Use PCs     : Yes
#> Found 24 associations (uncorrected p-values).
#> 
#> Combining p-values from multi-GWAS using fisher method
#> 
#> Overall, found 54 associations with (uncorrected) combined p-values).
#> 
#> ========== Analyzing chromosome/part 21
#> 
#> Performing GWAS with GLM over 1 chunks
#> Trait(s)    : BMI
#> Covariate(s): Age, Gender
#> Regression  : gaussian (predicted)
#> Use PCs     : Yes
#>   testing SNPs from rs2824974 to rs2828316
#> Found 18 associations (uncorrected p-values).
#> 
#> Preparing data for statgenGWAS...
#> Performing GWAS with statgenGWAS model
#> Trait(s)    : BMI
#> Covariate(s): Age, Gender
#> Use PCs     : Yes
#> Found 18 associations (uncorrected p-values).
#> 
#> Combining p-values from multi-GWAS using fisher method
#> 
#> Overall, found 35 associations with (uncorrected) combined p-values).
#> 
#> Finished, the following putative associations were found per part:
#>   Part 18: 54 associations found out of 657 markers
#>   Part 21: 35 associations found out of 343 markers
#> 
#> Overall, found 89 associations with (uncorrected) combined p-values).
    toy <- gwa(toy,response="BMI",covariates=c("Age","Gender"),
        pcs=FALSE,psig=0.05,methods=c("glm","statgen"))
#> 
#> Starting GWA analysis in 2 chunks with 2 methods: glm, statgen
#> 
#> ========== Analyzing chromosome/part 18
#> 
#> Performing GWAS with GLM over 1 chunks
#> Trait(s)    : BMI
#> Covariate(s): Age, Gender
#> Regression  : gaussian (predicted)
#> Use PCs     : No
#>   testing SNPs from rs1009238 to rs1568327
#> Found 25 associations (uncorrected p-values).
#> 
#> Preparing data for statgenGWAS...
#> Performing GWAS with statgenGWAS model
#> Trait(s)    : BMI
#> Covariate(s): Age, Gender
#> Use PCs     : No
#> Found 25 associations (uncorrected p-values).
#> 
#> Combining p-values from multi-GWAS using fisher method
#> 
#> Overall, found 52 associations with (uncorrected) combined p-values).
#> 
#> ========== Analyzing chromosome/part 21
#> 
#> Performing GWAS with GLM over 1 chunks
#> Trait(s)    : BMI
#> Covariate(s): Age, Gender
#> Regression  : gaussian (predicted)
#> Use PCs     : No
#>   testing SNPs from rs2824974 to rs2828316
#> Found 16 associations (uncorrected p-values).
#> 
#> Preparing data for statgenGWAS...
#> Performing GWAS with statgenGWAS model
#> Trait(s)    : BMI
#> Covariate(s): Age, Gender
#> Use PCs     : No
#> Found 16 associations (uncorrected p-values).
#> 
#> Combining p-values from multi-GWAS using fisher method
#> 
#> Overall, found 35 associations with (uncorrected) combined p-values).
#> 
#> Finished, the following putative associations were found per part:
#>   Part 18: 52 associations found out of 657 markers
#>   Part 21: 35 associations found out of 343 markers
#> 
#> Overall, found 87 associations with (uncorrected) combined p-values).
    
    listGwa(toy)
#> 1 -> Phenotype: BMI | Covariates: Age, Gender | PCs: 1
#>      Putative formula: BMI ~ Age + Gender + PC1 + [SNP]
#> 2 -> Phenotype: BMI | Covariates: Age, Gender | PCs: -
#>      Putative formula: BMI ~ Age + Gender + [SNP]
#> 
#> Note: The PC variables have putative names based on the number of PCs included in the model.
#> For example, instead of PC1, PC2, PC3 the actual PCs may be PC1, PC3, PC4
    
    prsiceOpts <- getDefaults("prsice")
    prsiceOpts$perm <- 100
    if (Sys.which("PRSice_linux") != "") {
        toy <- runPRS(base=toy,response="BMI",
            covariates=c("Age","Gender"),
            pcs=TRUE,prsiceOpts=prsiceOpts)
        toy <- runPRS(base=toy,response="BMI",
            covariates=c("Age","Gender"),
            pcs=FALSE,prsiceOpts=prsiceOpts)
    }
#> 
#> Starting PRS analysis with lassosum, prsice
#> 
#> ========== Running PRS analysis with lassosum
#> lassosum operations will be executed at /tmp/Rtmpwic9pB/lassosum_QYJSJ6461I
#> Preparing lassosum run at workspace directory: /tmp/Rtmpwic9pB/lassosum_QYJSJ6461I
#>   preparing summary statistics...
#>   preparing PLINK temporary files for target...
#> Preflight...
#> Writing FAM file to /tmp/Rtmpwic9pB/lassosum_QYJSJ6461I/plink_lassosum_target_VHSGZ1399P.fam 
#> Writing extended MAP file to /tmp/Rtmpwic9pB/lassosum_QYJSJ6461I/plink_lassosum_target_VHSGZ1399P.bim 
#> Writing BED file to /tmp/Rtmpwic9pB/lassosum_QYJSJ6461I/plink_lassosum_target_VHSGZ1399P.bed (SNP-major mode)
#>   copying linkage disequilibrium block files...
#> Preparation done!
#> Running lassosum PRS algorithm
#> Coordinating summary stats with reference panel...
#> Coordinating summary stats with test data...
#> Coordinating summary stats, reference panel, and test data...
#> Splitting genome by LD blocks ...
#> Running lassosum ...
#> s =  0.2 
#> s =  0.5 
#> s =  0.9 
#> Running lassosum with s=1...
#> Calculating polygenic scores ...
#> Running lassosum validation and PRS calculation
#> 100 out of 100 samples kept in pheno.
#> 100 out of 100 samples kept in covar.
#> Done! R^2 is: 0.756356747467697
#> 
#> ========== Running PRS analysis with prsice
#> prsice operations will be executed at /tmp/Rtmpwic9pB/prsice_HWUPU5469K
#> Preparing PRSice run at workspace directory: /tmp/Rtmpwic9pB/prsice_HWUPU5469K
#>   preparing summary statistics for the base file...
#>   preparing the covariates file...
#>   preparing the phenotype file...
#>   preparing PLINK temporary files - target...
#> Preflight...
#> Writing FAM file to /tmp/Rtmpwic9pB/prsice_HWUPU5469K/plink_prsice_YJTRP3784U.fam 
#> Writing extended MAP file to /tmp/Rtmpwic9pB/prsice_HWUPU5469K/plink_prsice_YJTRP3784U.bim 
#> Writing BED file to /tmp/Rtmpwic9pB/prsice_HWUPU5469K/plink_prsice_YJTRP3784U.bed (SNP-major mode)
#> Preparation done!
#> Running PRSice PRS algorithm in calculate mode
#> 
#> Executing:
#> Rscript /home/panos/.local/share/R/prsice-2.3.3/PRSice.R \
#>   --prsice /home/panos/.local/share/R/prsice-2.3.3/PRSice_linux \
#>   --base /tmp/Rtmpwic9pB/prsice_HWUPU5469K/base_YJTRP3784U \
#>   --beta \
#>   --binary-target F \
#>   --pheno /tmp/Rtmpwic9pB/prsice_HWUPU5469K/pheno_YJTRP3784U \
#>   --target /tmp/Rtmpwic9pB/prsice_HWUPU5469K/plink_prsice_YJTRP3784U \
#>   --cov /tmp/Rtmpwic9pB/prsice_HWUPU5469K/covar_YJTRP3784U \
#>   --clump-kb 250 \
#>   --clump-r2 0.1 \
#>   --clump-p 1 \
#>   --score avg \
#>   --seed 42 \
#>   --out prsice_out_YJTRP3784U \
#>   --print-snp \
#>   --perm 100
#> Done!
#> Adjusted PRSice R^2 is: 0.708854
#> Full model R^2 is: 0.867266
#> Null model R^2 is: 0.158412
#> 
#> Running PRS algorithms finished! Processing the results...
#> Done!
#> 
#> Starting PRS analysis with lassosum, prsice
#> 
#> ========== Running PRS analysis with lassosum
#> lassosum operations will be executed at /tmp/Rtmpwic9pB/lassosum_EWZRO6903A
#> Preparing lassosum run at workspace directory: /tmp/Rtmpwic9pB/lassosum_EWZRO6903A
#>   preparing summary statistics...
#>   preparing PLINK temporary files for target...
#> Preflight...
#> Writing FAM file to /tmp/Rtmpwic9pB/lassosum_EWZRO6903A/plink_lassosum_target_JVIAQ8582L.fam 
#> Writing extended MAP file to /tmp/Rtmpwic9pB/lassosum_EWZRO6903A/plink_lassosum_target_JVIAQ8582L.bim 
#> Writing BED file to /tmp/Rtmpwic9pB/lassosum_EWZRO6903A/plink_lassosum_target_JVIAQ8582L.bed (SNP-major mode)
#>   copying linkage disequilibrium block files...
#> Preparation done!
#> Running lassosum PRS algorithm
#> Coordinating summary stats with reference panel...
#> Coordinating summary stats with test data...
#> Coordinating summary stats, reference panel, and test data...
#> Splitting genome by LD blocks ...
#> Running lassosum ...
#> s =  0.2 
#> s =  0.5 
#> s =  0.9 
#> Running lassosum with s=1...
#> Calculating polygenic scores ...
#> Running lassosum validation and PRS calculation
#> 100 out of 100 samples kept in pheno.
#> 100 out of 100 samples kept in covar.
#> Done! R^2 is: 0.757420227733504
#> 
#> ========== Running PRS analysis with prsice
#> prsice operations will be executed at /tmp/Rtmpwic9pB/prsice_SAKPN7219U
#> Preparing PRSice run at workspace directory: /tmp/Rtmpwic9pB/prsice_SAKPN7219U
#>   preparing summary statistics for the base file...
#>   preparing the covariates file...
#>   preparing the phenotype file...
#>   preparing PLINK temporary files - target...
#> Preflight...
#> Writing FAM file to /tmp/Rtmpwic9pB/prsice_SAKPN7219U/plink_prsice_WEHUS1327G.fam 
#> Writing extended MAP file to /tmp/Rtmpwic9pB/prsice_SAKPN7219U/plink_prsice_WEHUS1327G.bim 
#> Writing BED file to /tmp/Rtmpwic9pB/prsice_SAKPN7219U/plink_prsice_WEHUS1327G.bed (SNP-major mode)
#> Preparation done!
#> Running PRSice PRS algorithm in calculate mode
#> 
#> Executing:
#> Rscript /home/panos/.local/share/R/prsice-2.3.3/PRSice.R \
#>   --prsice /home/panos/.local/share/R/prsice-2.3.3/PRSice_linux \
#>   --base /tmp/Rtmpwic9pB/prsice_SAKPN7219U/base_WEHUS1327G \
#>   --beta \
#>   --binary-target F \
#>   --pheno /tmp/Rtmpwic9pB/prsice_SAKPN7219U/pheno_WEHUS1327G \
#>   --target /tmp/Rtmpwic9pB/prsice_SAKPN7219U/plink_prsice_WEHUS1327G \
#>   --cov /tmp/Rtmpwic9pB/prsice_SAKPN7219U/covar_WEHUS1327G \
#>   --clump-kb 250 \
#>   --clump-r2 0.1 \
#>   --clump-p 1 \
#>   --score avg \
#>   --seed 42 \
#>   --out prsice_out_WEHUS1327G \
#>   --print-snp \
#>   --perm 100
#> Done!
#> Adjusted PRSice R^2 is: 0.702364
#> Full model R^2 is: 0.857239
#> Null model R^2 is: 0.154874
#> 
#> Running PRS algorithms finished! Processing the results...
#> Done!
        
    listPrs(toy)
#> 1 -> Phenotype: BMI | Covariates: Age, Gender | PCs: 1
#>      Putative formula: BMI ~ Age + Gender + PC1 + [PRS]
#> 2 -> Phenotype: BMI | Covariates: Age, Gender | PCs: -
#>      Putative formula: BMI ~ Age + Gender + [PRS]
#> 
#> Note: The PC variables have putative names based on the number of PCs included in the model.
#> For example, instead of PC1, PC2, PC3 the actual PCs may be PC1, PC3, PC4