Merge GWASExperiments from different populations

This function attempts to merge two GWASExperiments with various goals, such as merging similar populations with common phenotypic and/or anthropometric data to increase statistical power. See Details for a more rigorous explanation of the merging process.

Usage

mergeGWAS(gwe1, gwe2, output = c("common", "all"), 
        mafResolve = NULL, gdsfile = NULL, writegds = FALSE)

Arguments

gwe1

A GWASExperiment object from the first dataset.

gwe2

A GWASExperiment object from the second dataset.

output

One of "common" (default) to operate on common SNPs between gwe1 and gwe2 or "all" to operate on all SNPs. SNPs unique to each dataset are returned with missing genotypes in the merged set.

mafResolve

How would alleles that cannot be resolved just by strand fliping be resolved? Based on the frequency in one of the two datasets (1 or 2) or from online resources (dbSNP, population studies etc.). The later is the default (NULL).

gdsfile

path to the new GDS file. If NULL, a temporary file will be written (unadvised if writegds=TRUE!).

writegds

write the new GDS file? Defaults to FALSE. It can be written anytime with the GWASExperiment2GDS function.

Value

A new GWASExperiment object with merged genotypes, samples and phenotypes.

Details

The merging of GWASExperiment objects is performed based on the following steps and assumptions. Firstly, the two objects are checked for genome version compatibility. If they are not compatible the process does not start. You should consider using the GWASExperimentLiftOver function that can potentially render the two datasets compatible. Also, both datasets should have a common reference SNP naming. It is not possible to merge a dataset where the markers are named e.g. by Illumina IDs and a dataset where the markers are named with dbSNP (rs) IDs.

If the two datasets are compatible in terms of genome version and coordinates, the common SNPs are then identified between the two datasets. As these must point to the same information (location, minor and major alleles, same strand), the process then checks these attributes and tries to resolve any discrepancies.

The first check is whether non identical SNPs in terms of alleles can be solved with strand-fliping. Those that cannot be resolved with strand-fliping may not point to the same risk allele. This is resolved either by using one dataset as a reference (mafResolve=1 or 2) or even better, by querying online resources (e.g. Ensembl and dbSNP) to identify the true minor and major alleles based on summarized frequencies from population studies (there should be an internet connection for this). SNPs whose alleles cannot be resolved by any means (strand-flip, online MAF resolve) are completely dropped.

If the common SNPs between the two objects continue not being identical, it probably means that there are location/position mismatches potentially derived from older annotations. These can only be resolved online. If after this there are any remaining non-resolvable SNPs, they are dropped. Then, the genotypes of common SNPs are merged. The genotypes of the alleles that were flipped are also switched.

Finally, the common annotation elements and genotypes are combined. If output=="common" the process stops. If output="all", the unique elements of each dataset are also combined with the common. The genotypes of SNPs unique to each dataset are set to missing.

With respect to the allele resolving of the common SNPs that cannot be resolved with strand-flipping online:

SNPs are queried with biomaRt to resolve allele frequencies. Genome version is not important at this point as only alleles are used and IDs.

If minor allele is available from Ensembl, then the flip decision is based on it. If not (e.g. outdated/merged SNPs, dbSNP is queried with package rsnps and fill the Ensembl result.

Some SNPs may have been removed from dbSNP. Nothing can be done. Flip for these will be based on the majority of the rest.

Some found SNPs may not have information on MAF and minor alleles. dbSNP is requeried with rsnps and filling missing info is retried. Again, some will not be resolved and will be flipped according to majority.

Author

Panagiotis Moulos

Examples

    data(toy,package="prisma")
    
    toy1 <- toy[1:800,1:50]
    toy2 <- toy[201:1000,51:100]
    
    toym <- mergeGWAS(toy1,toy2)
#> Now trying to merge features (markers, SNPs)
#>   Found 600 common SNPs
#>   Checking common SNPs for reversed strands or alleles
#>   Found 0 SNPs with no mathcing alleles
#>   Found 0 SNPs fixable by strand-flip
#>   Checking for non-resolvable alleles
#>   Merging features
#> Now trying to merge samples
#> Now trying to construct unified genotypes
#>   Merging genotypes
#> Done!
#>  
    dim(toym) # 600 100
#> [1] 600 100
    
    toym <- mergeGWAS(toy1,toy2,output="all")
#> Now trying to merge features (markers, SNPs)
#>   Found 600 common SNPs
#>   Checking common SNPs for reversed strands or alleles
#>   Found 0 SNPs with no mathcing alleles
#>   Found 0 SNPs fixable by strand-flip
#>   Checking for non-resolvable alleles
#>   Retrieving SNPs specific to each dataset
#>   Merging features
#> Now trying to merge samples
#> Now trying to construct unified genotypes
#>   Merging genotypes
#> coercing object of mode  numeric  to SnpMatrix
#> object has no names - using numeric order for row/column names
#> coercing object of mode  numeric  to SnpMatrix
#> object has no names - using numeric order for row/column names
#> Done!
#>  
    dim(toym) # 1000 100
#> [1] 1000  100