Enrichment of PGS Catalog retrieved scores

This function uses three distinct tests to assess normality of a numeric vector, namely the Shapiro-Wilk test, the Kolmogorov Smirnov test against the normal distribution (y="pnorm" in the ks.test function) and the Jarque-Bera test implemented in the package tseries. If the latter is not present, the test is not available.

Usage

normalityCheck(x, pval = 0.05, tests = c("sw", "ks", "jb"),
        lower = 30, combine = c("fisher", "simes", "max", "min"), 
        nsample = 1000)

Arguments

x

a numeric vector to be tested for normality.

pval

p-value cutoff for the test significance.

tests

one or more of "sw" (Shapiro-Wilk), "ks" (Kolmogorov-Smirnov) and "jb" (Jarque-Bera).

lower

a lower threshold for sample size, below which the selected tests will be executed without resampling. See also Details.

combine

p-value combination if multiple tests are selected. One of "fisher", "simes", "max", "min". See also Details.

nsample

Number of resampling iterations when the sample size is larger than lower. Defaults to 1000.

Value

A named list with members statistics, the actual values of the statistics used to assess normality, pvalues, the respective p-values, "pval", the p-value threshold of the tests (same as input p, normal, a logical indicating the output of the test.)

Details

Regarding the lower argument, it represents a length threshold of the input vector (aka number of observations) above which the normality is assessed using resampling of x where each sample has length lower. The final outcome is then determined by the p = 100 * (1 - 2*pval) quantile of the nsample length Shapiro-Wilk p-value distribution created after nsample resampling iterations. If p >= pval, x can be considered to follow the normal distribution.

Regarding the combine argument, it determines how the outcomes of multiple tests will be combined. When "fisher" or "simes", then the Fisher or Simes method for p-value combination is used. For details on these two methods please refer to PMID: 25452340 and the Bioconductor package metaseqR2. When "min" or "max", then the minimum (more lose) or the maximum (stricter) p-value out of three tests is considered to determine the outcome.

Author

Panagiotis Moulos

Examples

x <- rnorm(20)

o1 <- normalityCheck(x)
#> Sample size 20 <= 30. Will execute combined normality checks using requested tests.
o2 <- normalityCheck(x,combine="simes")
#> Sample size 20 <= 30. Will execute combined normality checks using requested tests.
# Two instead of 3 tests
o3 <- normalityCheck(x,tests=c("sw","jb"),combine="simes")
#> Sample size 20 <= 30. Will execute combined normality checks using requested tests.