Functions to perform PRS cross-validation evaluations

These functions perform cross-validation regressions given either the output of prisma for full evaluation (prismaCrossValidate) or one PRS candidate data frame with SNPs in proper format (prsCrossValidate). Several metrics are returned for inspection and evaluation and are all used in the PRISMA HTML report.

Usage

prismaCrossValidate(prismaOut, gwe, response, 
        covariates = NULL, pcs = FALSE, 
        leaveOut = seq(0.05, 0.5, 0.05), times = 10,
        prsCalc = c("avg", "sum", "std"), family = NULL, 
        rc = NULL, ...)
    
    prsCrossValidate(snpSelection, gwe, response, 
        covariates = NULL, pcs = FALSE, leaveOut = 0.05,
        times = 10, prsCalc = c("avg", "sum", "std"),
        family = NULL, rc = NULL, ...)
    
    applyPRS(snpSelection, gwe, response, 
        covariates = NULL, pcs = FALSE, minSnps = 2,
        prsCalc = c("avg", "sum", "std"), times = 10,
        family = NULL, rc = NULL,...)

Arguments

prismaOut

an output from the link{prisma} function containing the results of an iterative PRS extraction.

gwe

a GWASExperiment object (same as the one used to generate prismaOut).

snpSelection

A data frame with candidate SNPs, the output, or part of the output of aggregatePrsMarkers, or one of the list members of the output of selectPrs.

response

The trait for which a PRS is sought to be extracted. It should be a character of length 1 and exist in the phenotypes of gwe.

covariates

The covariates to include in regression models. It should be a character vector of traits all existing in the phenotypes of gwe.

leaveOut

The fraction (percentage) of samples to leave-out in a cross-validation run. This can be a vector of multiple fractions for prismaCrossValidate.

times

For prismaCrossValidate, how many cross-validations to perform. For applyPRS, the number of SNP randomizations to perform if summary statistics are present to gwe. Defaults to 10. See also Details.

prsCalc

How the final PRS is calculated from the data frame of the PRS candidate SNPs? One of "avg" (default), "sum" or "std". See also Details.

family

Regression family of functions, e.g. "gaussian" for classical linear regression or "binomial" for logistic regression. See also gwa. Defaults to NULL for trying to set automatically.

pcs

Inlcude PCs in the regression model? Defaults to FALSE. See also Details.

minSnps

The minimum number of SNPs present in gwe in order to continue with PRS application using the genotypes of gwe. Defaults to 2.

rc

The fraction of available cores to use for parallel calculations. Default is NULL for no parallelization.

...

Further arguments passed to regression functions, like glm.

Value

For prismaCrossValidate, a list of cross-validation metrics, for each GWA method used, for each PRS candidate and for each leave-out percentage. The list is nested, with the first level being the GWA method, the second level being the SNP candidates (named after the number of SNPs in the PRS candidate) and the third level being the leave-out percentage of samples in cross-validation. The actual value of the third level is a data frame with times rows, containing the metrics collected after each cross-validation, leaving out leaveOut[i] fraction of samples, with i=1:ntimes. This data frame is the output of prsCrossValidate. For the collected metrics, see Details. For applyPRS a list with three or four members:

metrics: The same as the cross validation metrics - see Details.

values: A data frame with the following columns:

raw_pheno: The phenotype which the PRS contributes to its explanation

red_pheno: The predicted phenotype in a regression model without the PRS as a covariate

full_pheno: The predicted phenotype in a regression model including the PRS as a covariate

prs: The PRS calculated according to prsCalc

plots: A list of two ggplot objects:

scatter: A scatterplot of raw_pheno against prs

hist: A histogram of prs

permutations: This is available only when times > 0 and only when summary statistics (effects slot) are present in the gwe GWASExperiment object.

metrics: A matrics of evaluation metrics (same as the metrics upper level member) over times iterations (the number of rows)

bootp: A bootstrap p-value denoting the statistical significance of the tested PRS being non-random (essentially the number of times the PRS R^2 of the random PRS is larger than the original PRS R^2 divided by times)

Details

Regarding the prsCalc argument, this tells the CV functions how to calculate the PRS in the model given the genotypic data. The three available method reflect the three first ways of calculating the PRS in the PRSice2 software (https://www.prsice.info/step_by_step/) and are named accordingly (see the --score option in PRSice2).

Regarding the pcs argument, if TRUE, the PCs in gwe will be used. If gwe does not contain any PCs, a warning is issued and pcs is set to FALSE.

Regarding the times argument in the applyPRS function, this instructs the function to perform times random SNP selections from the input gwe and repeat the validation procedure through the calculation of evaluation metrics. This process is available only when summary statistics are present in gwe and is useful to test how a random PRS performs in explaining the examined phenotype. Normally, the metrics should be worse. A boostrap p-value is also calculated for some of the evaluation metrics.

Regarding the collected metrics, these are:

R^2 of the reduced regression model (reduced_r2): The R^2 of the reduced regression model, including all the covariates and PCs if requested, but not the PRS.

R^2 of the full regression model (full_r2): The R^2 of the full regression model including all the covariates and the PRS.

R^2 of the PRS (prs_r2): The adjusted R^2 of the PRS (essentially R^2 of full model minus the R^2 of the reduced model).

Statistical significance (p-value) of the PRS effect coefficient in the regression model (prs_pvalue).

R^2 of the reduced regression model (reduced_r2): The R^2 of the reduced regression model, including all the covariates and PCs if requested, but not the PRS.

RMSE (Root Mean Squared Error) of the reduced regression model (reduced_rmse): The RMSE of the reduced regression model, including all the covariates and PCs if requested, but not the PRS.

RMSE of the full regression model (full_rmse): The RMSE of the full regression model including all the covariates and the PRS.

MAE (Mean Absolute Error) of the reduced regression model (reduced_mae): The MAE of the reduced regression model, including all the covariates and PCs if requested, but not the PRS.

MAE of the full regression model (full_mae): The MAE of the full regression model including all the covariates and the PRS.

Correlation between observed and predicted phenotype values for the reduced regression model (reduced_pred_cor).

Correlation between observed and predicted phenotype values for the full regression model (full_pred_cor).

R^2 for the observed and predicted phenotype values for the reduced regression model (reduced_pred_r2).

R^2 for the observed and predicted phenotype values for the full regression model (full_pred_r2).

R^2 of the PRS for the observed and predicted phenotype values for the reduced regression model (prs_pred_r2).

Author

Panagiotis Moulos

Examples

    data(toy,package="prisma")
    data(prisma_out,package="prisma")
    
    response <- "BMI"
    covariates <- c("Case_Control","Gender","Age")
    
    cvm <- prismaCrossValidate(prismaOut,toy,response=response,
        covariates=covariates,leaveOut=c(0.2,0.4),times=5)
#> 
#> ################################################################
#> Running cross-validations for glm with 2 leave-out cases across 1 PRS candidates
#> ################################################################
#> 
#> ================================================================
#> -----> Cross-validating PRS candidate with 24 markers
#> ================================================================
#> 
#> ----- Leave-out percentage 20% -----
#> 
#> Regressing with GLM
#> Trait(s)    : BMI
#> Covariate(s): Case_Control, Gender, Age
#> Regression  : gaussian (predicted)
#> Use PCs     : No
#> 
#> ================================================================
#> -----> Cross-validation iteration 1
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 2
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 3
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 4
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 5
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ----- Leave-out percentage 40% -----
#> 
#> Regressing with GLM
#> Trait(s)    : BMI
#> Covariate(s): Case_Control, Gender, Age
#> Regression  : gaussian (predicted)
#> Use PCs     : No
#> 
#> ================================================================
#> -----> Cross-validation iteration 1
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 2
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 3
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 4
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 5
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ################################################################
#> Running cross-validations for statgen with 2 leave-out cases across 1 PRS candidates
#> ################################################################
#> 
#> ================================================================
#> -----> Cross-validating PRS candidate with 4 markers
#> ================================================================
#> 
#> ----- Leave-out percentage 20% -----
#> 
#> Regressing with GLM
#> Trait(s)    : BMI
#> Covariate(s): Case_Control, Gender, Age
#> Regression  : gaussian (predicted)
#> Use PCs     : No
#> 
#> ================================================================
#> -----> Cross-validation iteration 1
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 2
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 3
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 4
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 5
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ----- Leave-out percentage 40% -----
#> 
#> Regressing with GLM
#> Trait(s)    : BMI
#> Covariate(s): Case_Control, Gender, Age
#> Regression  : gaussian (predicted)
#> Use PCs     : No
#> 
#> ================================================================
#> -----> Cross-validation iteration 1
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 2
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 3
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 4
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 5
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ################################################################
#> Done! Use summarizeCvMetrics() and plotCvMetrics() to summarize and inspect the results.
#> ################################################################
    
    can <- getPrsCandidates(prismaOut,"glm",1)
    m <- prsCrossValidate(can,toy,response=response,
        covariates=covariates,leaveOut=0.2,times=5)
#> 
#> Regressing with GLM
#> Trait(s)    : BMI
#> Covariate(s): Case_Control, Gender, Age
#> Regression  : gaussian (predicted)
#> Use PCs     : No
#> 
#> ================================================================
#> -----> Cross-validation iteration 1
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 2
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 3
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 4
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model
#> 
#> ================================================================
#> -----> Cross-validation iteration 5
#> ================================================================
#> Partitioning dataset into 1 partitions by BMI
#> Creating the reduced model
#> Creating the full model