Functions to extract PRS candidates with various algorithms

These functions perform the PRS extraction given one or two (base and target) GWASExperiment objects. lassosumPRS is a wrapper around the lassosum PRS extraction algorithm while prsicePRS is a wrapper around the PRSice2 PRS extraction algorithm. The runPRS is a wrapper around the two others and makes sure that the proper objects are retured.

Usage

lassosumPRS(base, target = base, response, covariates = NULL,
        pcs = FALSE, wspace = NULL, anc = c("eur", "asn", "afr"),
        valid = c("auto", "stack", "split"), rc = NULL)
    
    PRS(gwe, snpSelection, type = c("avg", "sum", "std"))
    
    prsicePRS(base, target = base, response, covariates = NULL,
        pcs = FALSE, mode = c("calculate", "apply"), wspace = NULL,
        clump_kb = 250, clump_r2 = 0.1, clump_p = 1,
        score = c("avg", "sum", "std", "con-std"), perm = 10000,
        seed = 42, rc = NULL)
    
    runPRS(base, target = base, response, covariates = NULL,
        pcs = FALSE, methods = c("lassosum", "prsice"),
        lassosumOpts = getDefaults("lassosum"),
        prsiceOpts = getDefaults("prsice"), wspace = NULL, 
        rc = NULL)

Arguments

base

A GWASExperiment object with GWA tests performed and containing summary statistics apart from genotypes.

target

A GWASExperiment object with or without GWA tests, they are ignored. The PRS is calculated based on the target dataset.

response

The trait for which a PRS is sought to be extracted. It should be a character of length 1 and exist in the phenotypes of base and target.

covariates

The covariates to include in regression models. It should be a character vector of traits all existing in the phenotypes of base and target.

pcs

Inlcude PCs in the regression model? Defaults to FALSE.

mode

The mode to run PRSice2. One of calculating and extracting a PRS ("calculate", default) or applying a PRS given a GWASExperiment with existing prsbetas ("apply").

wspace

A path to a workspace directory for PRSice. If NULL, a temporary folder is used.

anc

One of the precalculated LD structures on lassosum, to be taken into account while extracting PRS. It can be "eur" (default) for European ancestry, "asn" for Asian or "afr" for African.

valid

Validation mode for lassosum PRS extraction. One of "auto" (default), "split" or "stack". See also the lassosum package.

gwe

A codeGWASExperiment object with the slot effects and/or prsbetas filled.

snpSelection

A data frame with candidate SNPs, the output, or part of the output of aggregatePrsMarkers, or one of the list members of the output of selectPrs.

type

Same as score in prsicePRS function (first three cases).

clump_kb

The genomic interval (in kbs) within which PRSice2 performs clumping. Defaults to 250.

clump_r2

The R^2 threshold to keep SNPs after clumping in PRSice2. Defaults to 0.1.

clump_p

The p-value threshold for SNPs to include in clumping. Defaults to 1.

score

How should PRSice2 calculate the PRS? One of "avg" (default), "sum", "std", "con-std". See also the manual of PRSice2 and Details.

perm

How many permutations should PRSice2 perform? Defaults to 10000.

seed

Seed for PRSice2 permutations, for reproducibility. Defaults to 42.

methods

The methods to use for PRS extraction, one or more of "prsice" and "lassosum".

lassosumOpts

List of options passed to lassosum (lassosumPRS). See also Details.

prsiceOpts

List of options passed to PRSice2 (prsicePRS). See also Details.

rc

The fraction of available cores to use for parallel calculations. Default is NULL for no parallelization.

Value

For lassosumPRS and prsicePRS, a data frame with the SNPs comprising the extracted PRS, compatible with the score files returned by the PGS Catalog and the candidates returned by the complete prisma pipeline. For runPRS, the input target GWASExperiment object with the prsbetas slots filled. Essentially the effect weights or ORs for the SNPs comprising the PRS.

Details

Regarding the score argument, it controls how PRSice2 calculates the PRS and creates regression models as well as how it reports it. Although typically it should be the same as the prsiceOpts$score argument in the main prisma pipeline (and it is when prisma is used), it is not obligatory as PRISMA recalculates the final PRS before reporting and validating. However, if not the same, the actual score distribution will differ much between PRSice2 results and PRISMA results.

Author

Panagiotis Moulos

Examples

    data(toy,package="prisma")
    gdsfile(toy) <- system.file(package="prisma","extdata","toy.gds")
    
    # Only 1 PC will be returned because of dataset size
    toy <- suppressWarnings(calcPcaCovar(toy,method="snprel"))
#> Performing PCA to select covariates to include in associations
#> Reading GDS file /tmp/RtmpezHjqb/temp_libpath372f48f1c2bc/prisma/extdata/toy.gds...
#> Performing LD pruning...
#> SNP pruning based on LD:
#> Excluding 0 SNP (non-autosomes or non-selection)
#> Excluding 38 SNPs (monomorphic: TRUE, MAF: NaN, missing rate: NaN)
#>     # of samples: 100
#>     # of SNPs: 962
#>     using 1 thread
#>     sliding window: 500,000 basepairs, Inf SNPs
#>     |LD| threshold: 0.2
#>     method: composite
#> Chromosome 18: 23.44%, 154/657
#> Chromosome 21: 21.28%, 73/343
#> 227 markers are selected in total.
#> LD pruning finished! 227 SNPs will be used for PCA.
#> 
#> Performing PCA...
#> Principal Component Analysis (PCA) on genotypes:
#> Excluding 773 SNPs (non-autosomes or non-selection)
#> Excluding 0 SNP (monomorphic: TRUE, MAF: NaN, missing rate: NaN)
#>     # of samples: 100
#>     # of SNPs: 227
#>     using 1 thread
#>     # of principal components: 32
#> PCA:    the sum of all selected genotypes (0,1,2) = 10268
#> CPU capabilities: Double-Precision SSE2
#> Mon Jun  6 20:20:29 2022    (internal increment: 13024)
#> [..................................................]  0%, ETC: ---        [==================================================] 100%, completed, 1s
#> Mon Jun  6 20:20:30 2022    Begin (eigenvalues and eigenvectors)
#> Mon Jun  6 20:20:30 2022    Done.
#> Selecting number of PCs to include
#>   with Tracy-Widom test for significant eigenvalues
#>     selected 1 PCs
    
    toy <- gwa(toy,response="BMI",covariates=c("Age","Gender"),
        pcs=TRUE,psig=0.05,methods=c("glm","statgen"))
#> 
#> Starting GWA analysis in 2 chunks with 2 methods: glm, statgen
#> 
#> ========== Analyzing chromosome/part 18
#> 
#> Performing GWAS with GLM over 1 chunks
#> Trait(s)    : BMI
#> Covariate(s): Age, Gender
#> Regression  : gaussian (predicted)
#> Use PCs     : Yes
#>   testing SNPs from rs1009238 to rs1568327
#> Found 24 associations (uncorrected p-values).
#> 
#> Preparing data for statgenGWAS...
#> Performing GWAS with statgenGWAS model
#> Trait(s)    : BMI
#> Covariate(s): Age, Gender
#> Use PCs     : Yes
#> Found 24 associations (uncorrected p-values).
#> 
#> Combining p-values from multi-GWAS using fisher method
#> 
#> Overall, found 54 associations with (uncorrected) combined p-values).
#> 
#> ========== Analyzing chromosome/part 21
#> 
#> Performing GWAS with GLM over 1 chunks
#> Trait(s)    : BMI
#> Covariate(s): Age, Gender
#> Regression  : gaussian (predicted)
#> Use PCs     : Yes
#>   testing SNPs from rs2824974 to rs2828316
#> Found 18 associations (uncorrected p-values).
#> 
#> Preparing data for statgenGWAS...
#> Performing GWAS with statgenGWAS model
#> Trait(s)    : BMI
#> Covariate(s): Age, Gender
#> Use PCs     : Yes
#> Found 18 associations (uncorrected p-values).
#> 
#> Combining p-values from multi-GWAS using fisher method
#> 
#> Overall, found 35 associations with (uncorrected) combined p-values).
#> 
#> Finished, the following putative associations were found per part:
#>   Part 18: 54 associations found out of 657 markers
#>   Part 21: 35 associations found out of 343 markers
#> 
#> Overall, found 89 associations with (uncorrected) combined p-values).
    
    prsiceOpts <- getDefaults("prsice")
    prsiceOpts$perm <- 100
    
    if (Sys.which("PRSice_linux") != "")
        toy <- runPRS(base=toy,response="BMI",
            covariates=c("Age","Gender"),pcs=TRUE,
            prsiceOpts=prsiceOpts)
#> 
#> Starting PRS analysis with lassosum, prsice
#> 
#> ========== Running PRS analysis with lassosum
#> lassosum operations will be executed at /tmp/Rtmpxo9Dha/lassosum_WSJPC7148G
#> Preparing lassosum run at workspace directory: /tmp/Rtmpxo9Dha/lassosum_WSJPC7148G
#>   preparing summary statistics...
#>   preparing PLINK temporary files for target...
#> Preflight...
#> Writing FAM file to /tmp/Rtmpxo9Dha/lassosum_WSJPC7148G/plink_lassosum_target_EIRID5993K.fam 
#> Writing extended MAP file to /tmp/Rtmpxo9Dha/lassosum_WSJPC7148G/plink_lassosum_target_EIRID5993K.bim 
#> Writing BED file to /tmp/Rtmpxo9Dha/lassosum_WSJPC7148G/plink_lassosum_target_EIRID5993K.bed (SNP-major mode)
#>   copying linkage disequilibrium block files...
#> Preparation done!
#> Running lassosum PRS algorithm
#> Coordinating summary stats with reference panel...
#> Coordinating summary stats with test data...
#> Coordinating summary stats, reference panel, and test data...
#> Splitting genome by LD blocks ...
#> Running lassosum ...
#> s =  0.2 
#> s =  0.5 
#> s =  0.9 
#> Running lassosum with s=1...
#> Calculating polygenic scores ...
#> Running lassosum validation and PRS calculation
#> 100 out of 100 samples kept in pheno.
#> 100 out of 100 samples kept in covar.
#> Done! R^2 is: 0.756356747467697
#> 
#> ========== Running PRS analysis with prsice
#> prsice operations will be executed at /tmp/Rtmpxo9Dha/prsice_FKVJN3606M
#> Preparing PRSice run at workspace directory: /tmp/Rtmpxo9Dha/prsice_FKVJN3606M
#>   preparing summary statistics for the base file...
#>   preparing the covariates file...
#>   preparing the phenotype file...
#>   preparing PLINK temporary files - target...
#> Preflight...
#> Writing FAM file to /tmp/Rtmpxo9Dha/prsice_FKVJN3606M/plink_prsice_WOGET5330Y.fam 
#> Writing extended MAP file to /tmp/Rtmpxo9Dha/prsice_FKVJN3606M/plink_prsice_WOGET5330Y.bim 
#> Writing BED file to /tmp/Rtmpxo9Dha/prsice_FKVJN3606M/plink_prsice_WOGET5330Y.bed (SNP-major mode)
#> Preparation done!
#> Running PRSice PRS algorithm in calculate mode
#> 
#> Executing:
#> Rscript /home/panos/.local/share/R/prsice-2.3.3/PRSice.R \
#>   --prsice /home/panos/.local/share/R/prsice-2.3.3/PRSice_linux \
#>   --base /tmp/Rtmpxo9Dha/prsice_FKVJN3606M/base_WOGET5330Y \
#>   --beta \
#>   --binary-target F \
#>   --pheno /tmp/Rtmpxo9Dha/prsice_FKVJN3606M/pheno_WOGET5330Y \
#>   --target /tmp/Rtmpxo9Dha/prsice_FKVJN3606M/plink_prsice_WOGET5330Y \
#>   --cov /tmp/Rtmpxo9Dha/prsice_FKVJN3606M/covar_WOGET5330Y \
#>   --clump-kb 250 \
#>   --clump-r2 0.1 \
#>   --clump-p 1 \
#>   --score avg \
#>   --seed 42 \
#>   --out prsice_out_WOGET5330Y \
#>   --print-snp \
#>   --perm 100
#> Done!
#> Adjusted PRSice R^2 is: 0.708854
#> Full model R^2 is: 0.867266
#> Null model R^2 is: 0.158412
#> 
#> Running PRS algorithms finished! Processing the results...
#> Done!
    
    lprs <- lassosumPRS(base=toy,response="BMI",
        covariates=c("Age","Gender"),pcs=TRUE)
#> lassosum operations will be executed at /tmp/Rtmpxo9Dha/lassosum_OAORP9897W
#> Preparing lassosum run at workspace directory: /tmp/Rtmpxo9Dha/lassosum_OAORP9897W
#>   preparing summary statistics...
#>   preparing PLINK temporary files for target...
#> Preflight...
#> Writing FAM file to /tmp/Rtmpxo9Dha/lassosum_OAORP9897W/plink_lassosum_target_VDWTZ8527W.fam 
#> Writing extended MAP file to /tmp/Rtmpxo9Dha/lassosum_OAORP9897W/plink_lassosum_target_VDWTZ8527W.bim 
#> Writing BED file to /tmp/Rtmpxo9Dha/lassosum_OAORP9897W/plink_lassosum_target_VDWTZ8527W.bed (SNP-major mode)
#>   copying linkage disequilibrium block files...
#> Preparation done!
#> Running lassosum PRS algorithm
#> Coordinating summary stats with reference panel...
#> Coordinating summary stats with test data...
#> Coordinating summary stats, reference panel, and test data...
#> Splitting genome by LD blocks ...
#> Running lassosum ...
#> s =  0.2 
#> s =  0.5 
#> s =  0.9 
#> Running lassosum with s=1...
#> Calculating polygenic scores ...
#> Running lassosum validation and PRS calculation
#> 100 out of 100 samples kept in pheno.
#> 100 out of 100 samples kept in covar.
#> Done! R^2 is: 0.756356747467697
    
    if (Sys.which("PRSice_linux") != "")
        pprs <- prsicePRS(base=toy,response="BMI",
            covariates=c("Age","Gender"),pcs=TRUE,perm=100)
#> prsice operations will be executed at /tmp/Rtmpxo9Dha/prsice_AYASI0629U
#> Preparing PRSice run at workspace directory: /tmp/Rtmpxo9Dha/prsice_AYASI0629U
#>   preparing summary statistics for the base file...
#>   preparing the covariates file...
#>   preparing the phenotype file...
#>   preparing PLINK temporary files - target...
#> Preflight...
#> Writing FAM file to /tmp/Rtmpxo9Dha/prsice_AYASI0629U/plink_prsice_RNNVO2324P.fam 
#> Writing extended MAP file to /tmp/Rtmpxo9Dha/prsice_AYASI0629U/plink_prsice_RNNVO2324P.bim 
#> Writing BED file to /tmp/Rtmpxo9Dha/prsice_AYASI0629U/plink_prsice_RNNVO2324P.bed (SNP-major mode)
#> Preparation done!
#> Running PRSice PRS algorithm in calculate mode
#> 
#> Executing:
#> Rscript /home/panos/.local/share/R/prsice-2.3.3/PRSice.R \
#>   --prsice /home/panos/.local/share/R/prsice-2.3.3/PRSice_linux \
#>   --base /tmp/Rtmpxo9Dha/prsice_AYASI0629U/base_RNNVO2324P \
#>   --beta \
#>   --binary-target F \
#>   --pheno /tmp/Rtmpxo9Dha/prsice_AYASI0629U/pheno_RNNVO2324P \
#>   --target /tmp/Rtmpxo9Dha/prsice_AYASI0629U/plink_prsice_RNNVO2324P \
#>   --cov /tmp/Rtmpxo9Dha/prsice_AYASI0629U/covar_RNNVO2324P \
#>   --clump-kb 250 \
#>   --clump-r2 0.1 \
#>   --clump-p 1 \
#>   --score avg \
#>   --seed 42 \
#>   --out prsice_out_RNNVO2324P \
#>   --print-snp \
#>   --perm 100
#> Done!
#> Adjusted PRSice R^2 is: 0.708854
#> Full model R^2 is: 0.867266
#> Null model R^2 is: 0.158412