Calculate PRS evaluation metrics

This function fits a regression model given a number of SNPs from which a PRS is calculated, in an iterative manner. Then, several metrics for goodness of fit are calculated and returned.

Usage

prsRegressionMetrics(snpSelection, gwe, response, 
        covariates = NULL, pcs = FALSE, step = 10, family = NULL,
        rc = NULL, ...)

Arguments

snpSelection

A data frame with candidate SNPs, the output, or part of the output of aggregatePrsMarkers, or one of the list members of the output of selectPrs.

gwe

A GWASExperiment object (same as the one used to generate prismaOut).

response

The trait for which a PRS is sought to be extracted. It should be a character of length 1 and exist in the phenotypes of gwe.

covariates

The covariates to include in regression models. It should be a character vector of traits all existing in the phenotypes of gwe.

pcs

Include PCs from gwe in the regression models? Defaults to FALSE.

step

Increment step for the number of SNPs from snpSelection to include in iterative evaluation. Defaults to 10.

family

Regression family of functions, e.g. "gaussian" for classical linear regression or "binomial" for logistic regression. See also gwa. Defaults to NULL for trying to set automatically.

rc

The fraction of available cores to use for parallel calculations. Default is NULL for no parallelization.

...

Further arguments passed to glm regression functions.

Value

A data frame with regression fit metrics, containing the following columns:

n_snp: The number of SNPs in the PRS being evaluated

freq: The frequency of appearance for these SNPs

full_r2: The R^2 of the full regression model, i.e. the regression model including the covariates and the PRS

full_pvalue: The p-value of the full regression model

prs_pvalue: The p-value of the PRS coefficient in the model

full_aic: The AIC of the full regression model

reduced_r2: The R^2 of the reduced model, i.e. the regression model including only the covariates and excluding the PRS

prs_r2: The R^2 attributed to the PRS (full_r2 - reduced_r2)

reduced_pvalue: The p-value of the reduced regression model

reduced_aic: The AIC of the reduced regression model

prs_aic: The AIC of the PRS (full_aic - reduced_aic)

Author

Panagiotis Moulos

Examples

    data(toy,package="prisma")
    data(prisma_out,package="prisma")
    
    response <- "BMI"
    covariates <- c("Case_Control","Gender","Age")
    
    can <- getPrsCandidates(prismaOut,"glm",1)
    m <- prsRegressionMetrics(can,toy,response=response,
        covariates=covariates,pcs=FALSE)
#> 
#> Regressing with GLM
#> Trait(s)    : BMI
#> Covariate(s): Case_Control, Gender, Age
#> Regression  : gaussian (predicted)
#> Use PCs     : No
#> Creating the reduced model
#> Creating the null model
#> Testing full models
#>   testing PRS with SNPs from rs11876399 to rs2825190
#>   testing PRS with SNPs from rs11876399 to rs8093690
#>   testing PRS with SNPs from rs11876399 to rs9977496