GWA test p-value combination functions

This group of fuctions can be used to calculate combined p-values, where each p-value is typically derived from a genome-wide association test for a SNP or a group of SNPs against a phenotype.

Usage

combineHarmonic(p, w = NULL, zerofix = NULL)
    combineMaxp(p)
    combineMinp(p)
    combineSimes(p, zerofix = NULL)
    combineWeight(p, w = NULL, zerofix = NULL)
    
    fisherMethod(p, zerofix = NULL)

Arguments

p

a numeric vector with p-values or a matrix for fisherMethod.

w

a weights vector, must sum to 1.

zerofix

NULL (default) or a fixed numeric value between 0 and 1.

Value

A new p-value, derived from the combination of p.

Details

The combineHarmonic function combines a vector of p-values using the method of Wilson, 2019 https://www.pnas.org/content/116/4/1195 and user- or prisma- provided weights.

The combineMaxp, combineMinp functions combine a vector of p-values using the maximum p-value (equivalent to the interesection of lists of significant SNPs returned by several GWA tests) or using the minimum p-value (equivalent to the union of lists of significant SNPs returned by several GWA tests).

The combineSimes function combines a vector of p-values using the method of Simes, 1986 method.

The combineWeight function combines a vector of p-values using the PANDORA algorithm described by Moulos and Hatzis, 2015 (PMID: 25452340) and Fanidis and Moulos, 2020 (PMID: 32778872), which is based on weighted p-value combination based on simulations from real data.

The fisherMethod function combines a vector of p-values using the respective method of Fisher.

The argument zerofix is used to correct for the case of a p-value which is equal to 0 as a result of internal numerical and approximation procedures. When NULL, random numbers greater than 0 and less than or equal to 0.5 are used to multiply the offending p-values with the lowest provided non-zero p-value, maintaining thus a virtual order of significance, avoiding having the same p-values for two tests and assuming that all zero p-values represent extreme statistical significance. When a numeric between 0 and 1, this number is used for the above multiplication instead.

Author

Panagiotis Moulos

Examples

    w <- c(0.2,0.5,0.3)
    p <- runif(3)
    
    pha <- combineHarmonic(p,w)
    pma <- combineMaxp(p)
    pmi <- combineMinp(p)
    psi <- combineSimes(p)
    pwe <- combineWeight(p,w)
    
    pm <- matrix(runif(300),100,3)
    
    pha <- apply(pm,1,combineHarmonic,w)
    pma <- apply(pm,1,combineMaxp)
    pmi <- apply(pm,1,combineMinp)
    psi <- apply(pm,1,combineSimes)
    pwe <- apply(pm,1,combineWeight,w)
    
    pwf <- fisherMethod(pm)