Functions to split datasets in training and test samples

These functions are used to create training and testing splits to be used in the iterative PRS extraction process of PRISMA. They are losely based on the dataset partitoning functions of the R package caret and can also be used independently of the general pipeline.

Usage

createSplit(y, n = 1, frac = 0.5, ngrp = min(5, length(y)),
        replace = FALSE, out = c("index", "binary"), rc = NULL)
    
    partitionGWAS(obj, by, n = 1, frac = 0.5, 
        ngrp = min(5, ncol(obj)), replace = FALSE,
        out = c("index", "train", "ttboth"), rc = NULL)

Arguments

y

A vector of numeric values to be split to a "training" and a "test" set and return the respective indices.

obj

Instead of y above, a GWASExperiment object.

by

The phenotype in obj to be used for splitting. It should be a scalar character.

n

How many splits to create? Defaults to 1.

frac

The fraction of values to remain in the "training" set. Defaults to 0.5.

ngrp

In how many "virtual" groups should y be split? Controls the uniformity of values in the two resulting sets. See also Details.

replace

Sample values/samples with replacement? Defaults to FALSE.

out

For createSplit, either "index" (default) to return a list of length n with "training" index splits, or "binary" to return the same list but instead of indices, it contains a mask of 0s and 1s indicating which samples have been selected and which discarded in the "training" set. For partitionGWAS, either "index" (default and same as createSplitm but now the indices refer to the columns of obj), or "train" to return a GWASExperiment with the training samples, or "ttboth" to return a list of GWASExperiments. In the latter case, the list has named members with names train and test respectively and the value of n is ignored, returning only one split.

rc

The fraction of available cores to use for parallel calculations. Default is NULL for no parallelization.

Value

For createSplit, a list of length n with the splits in the form of vector indices or a binary mask with 1s the training samples and 0s the test samples. For partitionGWAS, a list of indices as above or a list of GWASExperiment objects to be used as training samples, or a list with the training and testing GWASExperiment objects.

Author

Panagiotis Moulos

Examples

    data(toy,package="prisma")
    
    trainTest <- partitionGWAS(toy,by="BMI",out="ttboth")
#> Partitioning dataset into 1 partitions by BMI
    
    y <- phenotypes(toy)$BMI
    z <- createSplit(y,frac=0.8,n=5)