Tracy-Widom test for selection of PCs

This function implements the Tracy-Widom statistic for the assessment of the most significant principal components of a Principal Component Analysis. It is implemented based on assumptions described in Zhao et al., PMID: 30507552

Usage

twTest(eigv, p = 0.05, tol = 1e-8)

Arguments

eigv

The eigen values calculated from a PCA. See also Details.

p

p-value cutoff for the test significance.

tol

A value that will replace eigen values which are smaller than this value. Required for internal calculations.

Value

A named list with members: tw containing the TW statistic for each eigen value, pval containing the p-values and index containing the index of the significant eigen values, aka PCs.

Details

Regarding the eigv input, it can be retrieved according to which method or package has been used for PCA. For example, if robust PCA has been used (package rrcov), the getEigenvalues function can be used to get the eigen values. If prcomp has been used, then eigv is the squared sdev component of the prcomp output.

Author

Panagiotis Moulos

Examples

# Matrix borrowed from prcomp man page
C <- chol(S <- toeplitz(0.9^(0:31)))
X <- matrix(rnorm(32000),1000,32)
Z <- X 

P <- prcomp(Z)
tw <- twTest(P$sdev^2)
#> Warning: No signficant PCs found! Will return the first one...
# tw$index provides the significant PCs
identical(tw$index,which(tw$pval<0.05)) # TRUE
#> [1] FALSE